% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_postcodes.R
\name{add_rest_of_gb}
\alias{add_rest_of_gb}
\title{add rest of gb}
\usage{
add_rest_of_gb(postcode_list, all_postcodes, level, label = "Rest Of GB")
}
\arguments{
\item{postcode_list}{Named list of ONS postcodes for each area}

\item{all_postcodes}{Character vector of all postcodes}

\item{level}{level at which postcode has been truncated}

\item{label}{label for Rest Of GB}
}
\value{
updated postcode list
}
\description{
add rest of gb to postcode list
}
