% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airfare.R
\docType{class}
\name{airfare}
\alias{airfare}
\alias{airfare-class}
\title{\linkS4class{airfare}}
\description{
\linkS4class{airfare}
}
\section{Slots}{

\describe{
\item{\code{carrier}}{\link[base]{character} scalar}

\item{\code{depart}}{\link[base]{character} scalar}

\item{\code{arrive}}{\link[base]{character} scalar}

\item{\code{mileage}}{\link[base]{double} scalar}

\item{\code{code}}{\link[base]{character} scalar}

\item{\code{basefare,tax,carrier_imposed,upgrade}}{\link[base]{double} scalar}
}}

