% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ana099.R
\name{ana.099.im}
\alias{ana.099.im}
\title{ANA Round Trip from US}
\usage{
ana.099.im(data, US, min. = 21L, max. = 45L, ...)
}
\arguments{
\item{data}{\link[base]{data.frame}, downloaded from \url{https://ana.099.im}}

\item{US}{\link[base]{character} scalar or \link[base]{vector},
US airport for departure/arrival}

\item{min., max.}{\link[base]{integer} scalar, minimum and maximum days
between two departure dates.}

\item{...}{..}
}
\description{
ANA Round Trip from US
}
\details{
Must download latest \code{.csv} file from \url{https://ana.099.im}.

Upper-right corner of table, download symbol.
}
\examples{
if (FALSE) {
(fl = list.files(path = '~/Downloads', pattern = paste0('^', Sys.Date(), '.*_export.csv'), 
  full.names = TRUE))
dim(ana <- read.csv(file = fl, header = TRUE))
# unique(c(ana$departure, ana$arrival))
ana.099.im(ana, US = c('JFK', 'IAD'))
ana.099.im(ana, US = c('SFO', 'LAX'))
ana.099.im(ana, US = c('IAH'))
ana.099.im(ana, US = c('SEA', 'YVR'))
ana.099.im(ana, US = c('MEX'))
}
}
