\name{apply.shrink.dataframe}
\alias{apply.shrink.dataframe}

\title{Apply functions to columns in a dataframe}

\description{
\code{apply.shrink} for a dataframe, different functions can be applied to different columns, one for each column (?).
}

\usage{
apply.shrink.dataframe(data, name.x, name.ind, FUNS = NULL, NA.rm = FALSE, resp.name = NULL, full.data.frame = FALSE, Set = NA, name.res, ...)
}

\arguments{
  \item{data}{Input dataframe}
  \item{name.x}{Input value columns}
  \item{name.ind}{Category columns}
  \item{FUNS}{Functions to apply}
  \item{NA.rm}{na-action, default FALSE}
  \item{resp.name}{???}
  \item{full.data.frame}{???}
  \item{Set}{?}
  \item{name.res}{User selected values for the results columns}
  \item{\dots}{Additional arguments (to what???}
}

\value{
Dataframe of variouse outcomes.
}

\note{
Needs elaboration, although \dots are included in the arguments list they don't seem to be used anywhere.
}

\seealso{
\code{\link{apply.shrink}}, \code{\link{merge}}
}

\keyword{manip}
