\name{Arrow}
\alias{Arrow}

\title{
Add arrow to plot.
}

\description{
Adds arrow to a plot, given base and point position along with half angle.
}

\usage{
Arrow(pos, angle = 15, col = 2)
}

\arguments{
  \item{pos}{List with components \code{lat, lon} of length 2 or more (only first two used).}
  \item{angle}{Angle (< 90) determining arrow sharpness}
  \item{col}{Arrow color, defaults to color 2 of palette in effect.}
}

\value{
Mostly used for plotting, but the function returns the arrow polygon invisibly.
}

\seealso{
\code{\link{geopolygon}}, \code{\link{invProj}}
}

\examples{
geoplot()
Arrow(list(lat=c(65,65.5),lon=c(-19, -18)),angle=60,col="brown")
Arrow(list(lat=c(65,65.5),lon=c(-19, -18)),angle=45,col="red")
Arrow(list(lat=c(65,65.5),lon=c(-19, -18)),angle=30,col="green")
Arrow(list(lat=c(65,65.5),lon=c(-19, -18)),angle=30,col="blue")
}

\keyword{aplot}
