% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.iata.R
\name{as.iata}
\alias{as.iata}
\title{One or More Air Travel Trips using IATA codes}
\usage{
as.iata(x)
}
\arguments{
\item{x}{\link[base]{character} scalar or \link[base]{vector}}
}
\value{
The function \code{\link[=as.iata]{as.iata()}} returns an object of S3 class \code{'iata'}, which is essentially a \link[base]{list} of \link[base]{integer} \link[base]{vector}s.
}
\description{
One or more air travel trips using International Air Transport Association (IATA) airport codes.
}
\examples{
as.iata('NRT-HNL-YVR')
as.iata('CTU-PEK-ICN-JFK, EWR-IAH-LIM')
}
