% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postal_sectors.R
\name{assign_sector_to_area}
\alias{assign_sector_to_area}
\title{Assign sector to group}
\usage{
assign_sector_to_area(intersect_tbl, cutoff = 0.9)
}
\arguments{
\item{intersect_tbl}{Intersect values created by find_area_sector_coverage}

\item{cutoff}{Minimum ratio of overlaps required for assignment}
}
\description{
Find coverage of reference postal sectors by group
}
