% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{bisland}
\alias{bisland}
\title{Icelandic coastline, hi-resolution}
\format{A data frame with 19841 observations on the following 2 variables.
\describe{ \item{lat}{a numeric vector} \item{lon}{a numeric
vector} }}
\description{
Icelandic coastline, hi-resolution, approx. 10 times that of \code{island}.
}
\keyword{datasets}

