% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{calc_center_zoom}
\alias{calc_center_zoom}
\title{Calculate center and zoom}
\usage{
calc_center_zoom(lat, long)
}
\arguments{
\item{lat}{Latitudes}

\item{long}{Longitiudes}
}
\value{
List containing center of plot and suggested zoom level
}
\description{
Calculate center and zoom for a mapbox plot
}
