% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode_intersect.R
\name{check_postcode_intersect}
\alias{check_postcode_intersect}
\title{Check postcode intersect}
\usage{
check_postcode_intersect(
  postcodes1,
  postcodes2,
  id_cols,
  postcode_col = "postcode"
)
}
\arguments{
\item{postcodes1}{Table containing 1st set of postcodes}

\item{postcodes2}{Table containing 2nd set of postcodes}

\item{id_cols}{Grouping columns within postcode tables}

\item{postcode_col}{Column containing postcodes}
}
\value{
Table containing postcode intersect by grouping column
}
\description{
Check similarity between two postcode tables
}
