% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_postcodes.R
\name{check_postcodes}
\alias{check_postcodes}
\title{Check postcodes}
\usage{
check_postcodes(postcodes, level)
}
\arguments{
\item{postcodes}{character vector of postcodes}

\item{level}{level at which postcode has been truncated}
}
\description{
Checks all potcodes matches regular expressions
}
