% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colps.R
\name{colps}
\alias{colps}
\title{Open a postscript device with the color scheme given by geoplotpalette.}
\usage{
colps(...)
}
\arguments{
\item{...}{The arguments to postscript are optional. See help postscript.}
}
\value{
No value returned.
}
\description{
The function starts a postscript device and the arguments are the same as
the arguments to postscript (height, width, file, bg etc).
}
\details{
<!--explain details here-->
}
\section{Side Effects}{
 A graphics device is opened (often a file).  It must
be closed again by dev.off()
}
\examples{
\dontrun{
colps(file="map1.ps",height=6,width=5)
geoplot(xlim=c(-28,-10),ylim=c(64,69))
geosymbols(data,z=data$value,circles=0.2,sqrt=T)
geopolygon(island,col="white")# paint white over the symbols
geolines(island) # that are inside the country.  (island)
dev.off()

# same example in a different way.
colps(file="map1.ps",height=6,width=5,bg="white")
geoplot(xlim=c(-28,-10),ylim=c(64,69))
geosymbols(data,z=data$value,circles=0.2,sqrt=T)
geopolygon(island,col=0)#col 0 is now white
geolines(island) # was transparent earlier
dev.off()
}
}
\seealso{
postscript, geoplotpalette,litir
}
\keyword{<!--Put}
\keyword{here-->}
\keyword{more}
\keyword{one}
\keyword{or}
\keyword{s-keyword}
\keyword{tags}

