\name{colsymbol}
\alias{colsymbol}

\title{Plot colored symbols}

\description{
Adds colored symbols in a variety of shapes to a geo-plot.
}

\usage{
colsymbol(lat, lon, z, circles, squares, rectangles, hbars, vbars, perbars, parbars, levels, nlevels, colors, white, n, maxcol, digits, label.location, labels, fill.circles, density, angle, rotate, minsym = "<", label.resolution = 0, col = 1, labels.only = F, open.circles, lwd, border = F, bordercol = 0)
}

\arguments{
  \item{lat}{Latitude}
  \item{lon}{Longitude}
  \item{z}{Value}
  \item{circles}{If not zero, use circles of this size. }
  \item{squares}{If not zero, use circles of this size}
  \item{rectangles}{If not zero, use circles of this size}
  \item{hbars}{If not zero, use circles of this size}
  \item{vbars}{If not zero, use circles of this size}
  \item{perbars}{If not zero, use circles of this size}
  \item{parbars}{If not zero, use circles of this size}
  \item{levels}{Levels used for determining symbols size}
  \item{nlevels}{Number of levels}
  \item{colors}{Colors to use}
  \item{white}{Logical, use white for lowest level if TRUE}
  \item{n}{Number of points used to make a circle (?)}
  \item{maxcol}{maxcolor?}
  \item{digits}{??}
  \item{label.location}{Where to put legend} 
  \item{labels}{Labels for legend}
  \item{fill.circles}{Should circles be filled?}
  \item{density}{Density of shading when applicable}
  \item{angle}{Slant of shading}
  \item{rotate}{Should text (??) be rotated?}
  \item{minsym}{minimum value for a symbol to be drawn?}
  \item{label.resolution}{Number of digits in label???}
  \item{col}{Colors to use}
  \item{labels.only}{TRUE when labels/legend is added in a sperate call}
  \item{open.circles}{Should circles be open??}
  \item{lwd}{Line width of symbols}
  \item{border}{Should a border be drawn on the symbol}
  \item{bordercol}{Color for border if drawn}
}

\value{
No value returned, utility lies in side effect off adding colored symbols to existing plot, generally used as internal function in geosymbols.
}

\note{
Needs further elaboration, see documentation for \code{geosymbols}.
}

\seealso{
Called by \code{\link{geosymbols}}, calls \code{\link{Proj}}, \code{\link{geolocator}}, \code{\link{labels_size}}, \code{\link{labels1}}, \code{\link{labels2}}, \code{\link{shading1}}, \code{\link{paint.window}}
}

\keyword{aplot}
