% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{COUNTRY.DEFAULT}
\alias{COUNTRY.DEFAULT}
\title{Default country for \code{geo}.}
\format{The format is: chr 'island'}
\description{
Default country, \code{island}, given as a character string.
}
\keyword{datasets}

