% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_postcodes.R
\name{create_plot_data}
\alias{create_plot_data}
\title{Create plot data}
\usage{
create_plot_data(postcodes, postcode_list, area_yaml, dist_breaks)
}
\arguments{
\item{postcodes}{All ONS postcodes}

\item{postcode_list}{Named list of ONS postcodes for each area}

\item{area_yaml}{Portion of YAML file used to generate postcode list}

\item{dist_breaks}{Breaks for calculating distance. Must have 'm' units}
}
\value{
Returns plot data in format for postcode app
}
\description{
Create data for plotting in area app
}
