% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{create_polygon_from_str}
\alias{create_polygon_from_str}
\title{Create polygon from str}
\usage{
create_polygon_from_str(coordinate_str, in_crs, out_crs = 4326)
}
\arguments{
\item{coordinate_str}{A string of the polygon points}

\item{in_crs}{CRS of the input points}

\item{out_crs}{CRS of the output polygon}
}
\value{
A SFC object containing the polygon
}
\description{
Creates a polygon from a string listing polgon points
}
