% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{create_str_from_polygon}
\alias{create_str_from_polygon}
\title{Create str from polygon}
\usage{
create_str_from_polygon(polygon)
}
\arguments{
\item{polygon}{A SFC object containing the polygon}
}
\value{
A list containing the coordinate string and the CRS
}
\description{
Creates a string of coordinates from an SFC polygon
}
