\name{currentarrows}
\alias{currentarrows}

\title{
Plot arrows and segments showing the size and direction of currents.
}
\description{
Plot arrows and segments showing the size and direction of currents.
}
\usage{
currentarrows(data, maxsize = 0.5, maxn, col = "blue", lwd = 2, arrowsize = 0.2, center = T)
}

\arguments{
  \item{data}{
Data in a list with components \code{lat} and \code{lon} with decimal degrees, and \code{current} with the current magnitude.
}
  \item{maxsize}{
Maximum current segment size.
}
  \item{maxn}{
Current given with \code{maxsize}, defaults to \code{max(data$current)}.
}
  \item{col}{
Color of current arrows and segments.
}
  \item{lwd}{
Line width of the segments showing current.
}
  \item{arrowsize}{
Arrow size.
}
  \item{center}{
Whether or not to center the arrow, defaults to \code{TRUE}.
}
}
\note{
Needs further checking and elaboration.
}

\keyword{aplot}
