\name{cut_box.1}
\alias{cut_box.1}

\title{Cut box ?}

\description{
Cut box in connection with overlaying a Lambert projected plot with a grid.
}

\usage{
cut_box.1(x, y, xb, yb)
}

\arguments{
  \item{x, y}{Longitude and latitude of gridlines}
  \item{xb, yb}{Longitude and latitude limits}
}

\value{
List with compoents:
  \item{x}{Longitudes??}
  \item{y}{Latitudes??}
  \item{ind}{Indices??}
}

\note{
Internal in \code{gridaxes.Lambert}, needs elaboration.
}

\seealso{
\code{\link{gridaxes.Lambert}}
}

\keyword{manip}
