\name{cut_box.2}
\alias{cut_box.2}

\title{Cut box type 2???}

\description{
Cut box type 2???.
}

\usage{
cut_box.2(x, y, xb, yb)
}

\arguments{
  \item{x}{Longtude ?}
  \item{y}{Latitude ?}
  \item{xb}{Limits of longitude?}
  \item{yb}{ Limits of latitude?}
}

\value{
List with components:
  \item{x, y}{Longitude and latitude of ???}
  \item{x1, y1}{Some other Longitude and latitude of ???}
}

\note{
Internal to \code{gridaxes.Lambert}, needs elaboration.
}

\seealso{
\code{\link{gridaxes.Lambert}}
}

\keyword{manip}
