% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{depthloc}
\alias{depthloc}
\title{Depth label locations.}
\format{A data frame with 9 observations on the following 3 variables.
\describe{ \item{lat}{a numeric vector} \item{lon}{a numeric
vector} \item{z}{a numeric vector} }}
\description{
Standard locations for the labels of depth contours
given in \code{\link{gbdypi}}.
}
\keyword{datasets}
