% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distGeo_.R
\name{distGeo_}
\alias{distGeo_}
\title{Distance on an Ellipsoid, as \link[stats]{dist} Object}
\usage{
distGeo_(x, Labels = dimnames(x)[[1L]], ...)
}
\arguments{
\item{x}{2-column \link[base]{numeric} \link[base]{matrix}, coordinates}

\item{Labels}{see \strong{Value} section of \link[stats]{dist}}

\item{...}{additional parameters, currently not in use}
}
\description{
..
}
\details{
Function \code{\link[=distGeo_]{distGeo_()}} ...
}
\note{
Function \code{\link[=distGeo_]{distGeo_()}} can be passed into argument \code{distfun} of \link[stats]{heatmap}.
}
