% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earnAA.R
\name{earnAA}
\alias{earnAA}
\title{Credit a Flight to American Airlines}
\usage{
earnAA(airfare)
}
\arguments{
\item{airfare}{an \linkS4class{airfare}}
}
\description{
Credit a Flight to American Airlines
}
\keyword{internal}
