% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earnAS.R
\name{earnAS}
\alias{earnAS}
\title{Credit a Flight to Alaska Airlines}
\usage{
earnAS(airfare)
}
\arguments{
\item{airfare}{an \linkS4class{airfare}}
}
\description{
Credit a Flight to Alaska Airlines
}
\keyword{internal}
