% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earnBA.R
\name{earnBA}
\alias{earnBA}
\title{Credit a Flight to British Airways}
\usage{
earnBA(airfare)
}
\arguments{
\item{airfare}{an \linkS4class{airfare}}
}
\description{
Credit a Flight to British Airways
}
\keyword{internal}
