% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earnOneWorld.R
\name{earnOneWorld}
\alias{earnOneWorld}
\alias{earnAA}
\alias{earnAS}
\alias{earnASold}
\alias{earnBA}
\alias{earnCX}
\alias{earnJAL}
\title{Credit a Flight to an Airline in oneworld Alliance}
\usage{
earnAA(x, creditcard = "citi")

earnAS(x)

earnASold(x)

earnBA(x)

earnCX(x)

earnJAL(x)
}
\arguments{
\item{x}{an \linkS4class{airfare}}

\item{creditcard}{slot of class \linkS4class{loyalty}}
}
\description{
Credit a Flight to an Airline in oneworld Alliance
}
\keyword{internal}
