\name{Elimcomp}
\alias{Elimcomp}

\title{Parameters manipulation ?}
\description{
Parameters manipulation?
}
\usage{
Elimcomp(parlist)
}

\arguments{
  \item{parlist}{List of parameters ?}
}

\value{
Manipulated (shortened ?) parameterlist
}

\note{
Needs elaboration, \code{selectedpar} warps around this function, where parameter names in \code{nonsetpar} come from is a mystery.
}

\seealso{
Called by \code{\link{geoplot}}, \code{\link{init}} and \code{\link{selectedpar}}.
}

\keyword{device}
