% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{faeroes}
\alias{faeroes}
\title{Coastline of Faroe Islands}
\format{A data frame with 2161 observations on the following 2 variables.
\describe{ \item{lat}{a numeric vector} \item{lon}{a numeric
vector} }}
\source{
Possibly from package \code{mapdata}.
}
\description{
Old version of the coastline of the Faroes, with legacy typo.
}
\seealso{
faroes
}
\keyword{datasets}
