% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{filter_postcodes_distance}
\alias{filter_postcodes_distance}
\title{Filter postcodes distance}
\usage{
filter_postcodes_distance(postcodes, point, max_distance)
}
\arguments{
\item{postcodes}{sf object containing postcodes}

\item{point}{point from which to measure distance}

\item{max_distance}{unit object listing maximum distance}
}
\description{
Filter sf object by distance from point
}
