% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{filter_postcodes_polygon}
\alias{filter_postcodes_polygon}
\title{Filter postcodes polygon}
\usage{
filter_postcodes_polygon(postcodes, polygon)
}
\arguments{
\item{postcodes}{Simple feature object containing postcodes}

\item{polygon}{Simple feature collection containing single polygon}
}
\description{
Filter postcodes by their location within a polygon
}
