% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{filter_postcodes_regx}
\alias{filter_postcodes_regx}
\title{Filter postcodes regx}
\usage{
filter_postcodes_regx(postcodes, regex)
}
\arguments{
\item{postcodes}{Simple feature object containing postcodes}

\item{regx}{Regx pattern to search for in column}
}
\description{
Filter postcodes using regx
}
