\name{find.hnit}
\alias{find.hnit}

\title{Find coordinate(s) ??}

\description{
Find coordinates with some sort of interpolation (??).
}

\usage{
find.hnit(pt, poly)
}

\arguments{
  \item{pt}{Point(s) ??}
  \item{poly}{Polygon (??)}
}

\value{
Returns list with components:
  \item{x, y }{of coordinates}
}

\note{ ~~further notes~~ 
Needs elaboration.
}

\seealso{
Called by \code{\link{findcut}}
}

\keyword{manip}
