% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postal_sectors.R
\name{find_area_sector_coverage}
\alias{find_area_sector_coverage}
\title{Find group sector coverage}
\usage{
find_area_sector_coverage(
  query_postcodes,
  query_areas,
  ref_postcodes,
  unique = TRUE
)
}
\arguments{
\item{query_postcodes}{Vector of unique query postcodes}

\item{query_areas}{Vector of areas to which query postcodes belong}

\item{ref_postcodes}{Vector of unique reference postcodes}
}
\description{
Find coverage of reference postal sectors by group
}
