% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{find_postcode_distance_bins}
\alias{find_postcode_distance_bins}
\title{Find proximal postcodes}
\usage{
find_postcode_distance_bins(postcodes, ref_postcodes, dist_breaks)
}
\arguments{
\item{postcodes}{Query postcodes for which distances will be calculated}

\item{ref_postcodes}{Reference postcodes from which distance will be calculated}

\item{dist_breaks}{Distance breaks used to generate distance bins}
}
\description{
Find distance range of postcodes to reference postcodes
}
