% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postal_sectors.R
\name{find_sector_intersect}
\alias{find_sector_intersect}
\title{Find sector list intersect}
\usage{
find_sector_intersect(query_postcodes, ref_postcodes, query_only = TRUE)
}
\arguments{
\item{query_postcodes}{Vector of unique query postcodes}

\item{query_only}{Only include query postal sectors in output}

\item{ref_sector_list}{Vector of unique reference postcodes}
}
\description{
Find intersect of postcodes split by postal sector
}
