% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postal_sectors.R
\name{find_sector_list_intersect}
\alias{find_sector_list_intersect}
\title{Find sector intersect}
\usage{
find_sector_list_intersect(
  query_sector_list,
  ref_sector_list,
  query_only = TRUE
)

find_sector_list_intersect(
  query_sector_list,
  ref_sector_list,
  query_only = TRUE
)
}
\arguments{
\item{query_sector_list}{List of query postcodes split by postal sector}

\item{ref_sector_list}{List of reference postcodes split by postal sector}

\item{query_only}{Only include query postal sectors in output}

\item{ref_postcodes}{List of reference postcodes split by postal sector}
}
\description{
Find intersect of postcodes by postal sector

Find intersect of list of postcodes split by postal sector
}
