% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_postcodes.R
\name{format_complete_postcodes}
\alias{format_complete_postcodes}
\title{format complete postcodes}
\usage{
format_complete_postcodes(
  postcodes,
  to_upper = FALSE,
  replace_malformed = FALSE,
  sort = FALSE
)
}
\arguments{
\item{postcodes}{character vector of postcodes}

\item{to_upper}{convert postcodes to uppercase}

\item{replace_malformed}{replace malformed postcodes with NA instead of raising error}
}
\value{
formatted postcodes with single space between outward and inward code
}
\description{
format complete postcodes with a single space between outward and inward code
}
