% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{gbdypi.100}
\alias{gbdypi.100}
\title{GEBCO 100 m isobath in the N-Atlantic.}
\format{A data frame with 19190 observations on the following 2 variables.
\describe{ \item{lat}{a numeric vector} \item{lon}{a numeric
vector} }}
\description{
GEBCO 100 m isobath in the N-Atlantic.
}
\examples{

data(gbdypi.100)
## maybe str(gbdypi.100) ; plot(gbdypi.100) ...

}
\keyword{datasets}

