% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{gbdypi.1000}
\alias{gbdypi.1000}
\title{GEBCO 1000 m isobath in the N-Atlantic.}
\format{A data frame with 14869 observations on the following 2 variables.
\describe{ \item{lat}{a numeric vector} \item{lon}{a numeric
vector} }}
\description{
GEBCO 1000 m isobath in the N-Atlantic.
}
\examples{

data(gbdypi.1000)
## maybe str(gbdypi.1000) ; plot(gbdypi.1000) ...

}
\keyword{datasets}

