% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{gbdypi.200}
\alias{gbdypi.200}
\title{GEBCO 200 m isobath in the N-Atlantic.}
\format{A data frame with 29665 observations on the following 2 variables.
\describe{ \item{lat}{a numeric vector} \item{lon}{a numeric
vector} }}
\description{
GEBCO 200 m isobath in the N-Atlantic.
}
\examples{

data(gbdypi.200)
## maybe str(gbdypi.200) ; plot(gbdypi.200) ...

}
\keyword{datasets}
