% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{gbdypi.400}
\alias{gbdypi.400}
\title{GEBCO 400 m isobath in the N-Atlantic.}
\format{A data frame with 7201 observations on the following 2 variables.
\describe{ \item{lat}{a numeric vector} \item{lon}{a numeric
vector} }}
\description{
GEBCO 400 m isobath in the N-Atlantic.
}
\keyword{datasets}

