% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{gbdypi.500}
\alias{gbdypi.500}
\title{GEBCO 500 m isobath in the N-Atlantic.}
\format{A data frame with 16679 observations on the following 2 variables.
\describe{ \item{lat}{a numeric vector} \item{lon}{a numeric
vector} }}
\description{
GEBCO 500 m isobath in the N-Atlantic.
}
\examples{

data(gbdypi.500)
## maybe str(gbdypi.500) ; plot(gbdypi.500) ...

}
\keyword{datasets}
