% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{gbdypi.800}
\alias{gbdypi.800}
\title{GEBCO 800 m isobath in the N-Atlantic.}
\format{A data frame with 3083 observations on the following 2 variables.
\describe{ \item{lat}{a numeric vector} \item{lon}{a numeric
vector} }}
\description{
GEBCO 800 m isobath in the N-Atlantic.
}
\examples{
data(gbdypi.800)
## maybe str(gbdypi.800) ; plot(gbdypi.800) ...
}
\keyword{datasets}

