% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{gbdypi}
\alias{gbdypi}
\title{Isobaths in the NE-Atlantic}
\format{The format is: List of 58 data frame of lat and lon giving
isobaths for the same number of depths ranging from 50 to 6000 m
at 100 m intervals in the range 200 to 5500 m.}
\source{
IBCAO and/or GEBCO
}
\description{
An list of isobaths from IBCAO and/or GEBCO for selected depths.
}
\keyword{datasets}
