% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{gbdypif.200}
\alias{gbdypif.200}
\title{GEBCO 200 m isobath around Iceland.}
\format{A data frame with 914 observations on the following 2 variables:
\describe{ \item{lat}{a numeric vector} \item{lon}{a numeric
vector} }}
\description{
Given in one polygon.
}
\examples{
data(gbdypif.200)
## maybe str(gbdypif.200) ; plot(gbdypif.200) ...
}
\keyword{datasets}

