% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{gbdypif.400}
\alias{gbdypif.400}
\title{GEBCO 400 m isobath around Iceland.}
\format{A list of 4 data.frames of the following 2 variables:
\describe{ \item{lat}{a numeric vector} \item{lon}{a numeric
vector} }}
\description{
Given as 4 polygons.
}
\examples{
data(gbdypif.400)
## maybe str(gbdypif.400) ; plot(gbdypif.400) ...
}
\keyword{datasets}

