% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{gbdypif.500}
\alias{gbdypif.500}
\title{GEBCO 500 m isobath around Iceland.}
\format{A list of 2 data.frames of the following 2 variables:
\describe{ \item{lat}{a numeric vector} \item{lon}{a numeric
vector} }}
\description{
Given as 2 polygons.
}
\examples{

data(gbdypif.500)
## maybe str(gbdypif.500) ; plot(gbdypif.500) ...

}
\keyword{datasets}

