% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_.R
\name{gc_}
\alias{gc_}
\title{Great Circle as \link[sp]{SpatialLines} Object}
\usage{
gc_(x)
}
\arguments{
\item{x}{\link[sp]{SpatialPoints} object}
}
\value{
Function \link{gc_} returns an \link[sf]{sf} object.
}
\description{
..
}
