% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_postcodes.R
\name{generate_postcode_data}
\alias{generate_postcode_data}
\title{Generate postcode data}
\usage{
generate_postcode_data(yaml_path)
}
\arguments{
\item{yaml_path}{Path to YAML defining areas}
}
\value{
A list containing postcode tables and postcode plot data
}
\description{
Function generates postcode tables and postcode plot data
}
