% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_postcodes.R
\name{generate_postcode_list}
\alias{generate_postcode_list}
\title{Generate postcode list}
\usage{
generate_postcode_list(postcodes, definitions, operations, areas, level)
}
\arguments{
\item{postcodes}{ONS postcodes}

\item{definitions}{A list of definitions}

\item{operations}{A list of operations}

\item{areas}{A character vector of areas}

\item{level}{Postcode level}
}
\value{
A list of postcodes
}
\description{
Generate list of postcodes from supplied definitions and operations
}
