% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{generate_postcode_outputs}
\alias{generate_postcode_outputs}
\title{Generate postcode output}
\usage{
generate_postcode_outputs(postcodes, yaml_path, dist_breaks)
}
\arguments{
\item{postcodes}{ONS postcodes}

\item{yaml_path}{Path to YAML defining areas}

\item{dist_breaks}{Breaks for calculating distance. Must have 'm' units}
}
\description{
Function generates postcode tables and postcode plot data
}
