% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_postcodes.R
\name{generate_pull_data}
\alias{generate_pull_data}
\title{Generate postcode output}
\usage{
generate_pull_data(yaml_path)
}
\arguments{
\item{yaml_path}{Path to YAML defining areas}

\item{postcodes}{ONS postcodes}
}
\value{
A list containing postcode tables and postcode plot data
}
\description{
Function generates postcode tables and postcode plot data
}
