\name{geoaxis}
\alias{geoaxis}

\title{Setup axes for a geoplot}

\description{
Set up axes for a geoplot.
}

\usage{
geoaxis(side, pos, dist, dlat = 0.5, dlon = 1, csi = 0.12, cex = 0.7, inside = T, r = 1, ...)
}

\arguments{
  \item{side}{On which side should the axis be drawn?}
  \item{pos}{Positions of labels (?)}
  \item{dist}{Distance from plot of axes labels (?)}
  \item{dlat}{Latitude resolution of labels (?)}
  \item{dlon}{Longitude resolution of labels (?)}
  \item{csi}{Character size inches, default 0.12}
  \item{cex}{Character size expansion, default 0.7}
  \item{inside}{Whether or not stay inside ??????????}
  \item{r}{yet another setting ?????????}
  \item{\dots}{Additional arguments to \code{geotext}}
}

\value{
Adds labels to a geoplot, no value returned.
}

\note{
Needs further elaboration.
}

\seealso{
Calls \code{\link{geotext}}, called by \code{\link{init}}, commented out of \code{gridaxes.Lambert}.
}

\keyword{aplot}
