% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geodefine.R
\name{geodefine}
\alias{geodefine}
\title{Defines regions.}
\usage{
geodefine(nholes = 0)
}
\arguments{
\item{nholes}{The number of holes in data, number of regions - 1.}
}
\value{
A list of the points pointed out by the user with NA's between
regions.
}
\description{
locates points pointed out by users and defines regions.
}
\details{
Draws the regions on the plot.
}
\section{Side Effects}{
 Draws the regions on the plot.
}
\examples{
##   Push left mouse button to mark point, push middle button to
 ##   mark the end of a region.
}
\seealso{
\code{\link{geolocator}}.
}

