\name{geodezoom}
\alias{geodezoom}
\title{
  Restores a zoomed plot.
}
\description{
Geodezoom restores a plot zoomed with geodezoom.
}
\usage{
geodezoom()
}
\value{
none
}
\section{Side Effects}{
The limits of the current plot change back to its original
       size.
}
\seealso{
\code{\link{geoplot}},
\code{\link{geozoom}}.
}
\examples{
##    See examples in help(geozoom).
   
}
% docclass is function
% Converted by Sd2Rd version 43268.
