% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geointersect.R
\name{geointersect}
\alias{geointersect}
\title{Find intersection of 2 polygons}
\usage{
geointersect(data, border, in.or.out)
}
\arguments{
\item{data}{Polygon.}

\item{border}{Polygon to intersect with/complement from.}

\item{in.or.out}{Whether to take intersect of ‘x’ and ‘xb’ (0) 
or complement of ‘x’ in ‘xb’ (1). Default 0.}
}
\description{
Find intersection of 2 polygons
}
\seealso{
\code{\link{findcut}}, \code{\link{invProj}}, \code{\link{Proj}}
}
\keyword{logic}
\keyword{manip}

