% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IATA.R
\name{geom_worldmap}
\alias{geom_worldmap}
\title{Layer of World Map}
\usage{
geom_worldmap(...)
}
\arguments{
\item{...}{parameters of \link[ggplot2]{layer},
to be passed into \link[ggplot2]{geom_map}}
}
\description{
..
}
\examples{
# example code
library(ggplot2)
ggplot() + geom_worldmap() + xlim(c(-160, 175)) + ylim(c(-75, 80))
ggplot() + geom_worldmap() # does not work!!

}
\references{
\url{https://www.riinu.me/2022/02/world-map-ggplot2/}
}
