% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geoplotbwpalette.R
\name{geoplotbwpalette}
\alias{geoplotbwpalette}
\title{Black and white palette}
\usage{
geoplotbwpalette()
}
\value{
Sets the palette, no value.
}
\description{
Sets palette to 155 or so greys.
}
\note{
Redundant, deprecate?
}
\seealso{
Called by \code{\link{bwps}}.
}
\keyword{color}

