% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geozoom.R
\name{geozoom}
\alias{geozoom}
\title{Zoom into plots.}
\usage{
geozoom()
}
\value{
none.
}
\description{
Zoom into current geoplot, redraws current geoplot with area defined by
user.  The current plot can be restored by using geodezoom.
}
\section{Side Effects}{
 The last call to geoplot is recalled with different
borders for x and y.
}

\examples{

    geoplot()
    geozoom()
    # Click with mouse as when placing legend i.e. first place the
    # mouse where the upper left corner is supposed to be and click
    # once, then move to where the lower right corner is supposed to
    # be and also press once.
    geodezoom()
    # Return to previous plot, here geoplot().

}
\seealso{
\code{\link{geoplot}}, \code{\link{geodezoom}}.
}
