% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{gepco500}
\alias{gepco500}
\title{GEBCO 500 m isobath extending around the Faroes.}
\format{A data frame with 679 observations on the following 2 variables.
\describe{ \item{lat}{a numeric vector} \item{lon}{a numeric
vector} }}
\description{
GEBCO 500 m isobath extending around the Faroes.
}
\examples{

data(gepco500)
## maybe str(gepco500) ; plot(gepco500) ...

}
\keyword{datasets}
