% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_postcodes.R
\name{get_county_postcodes}
\alias{get_county_postcodes}
\title{Get counties postcodes}
\usage{
get_county_postcodes(postcodes, counties)
}
\arguments{
\item{postcodes}{sf object containing postcodes}

\item{counties}{A character vector of county codes}
}
\value{
sf object containing postcodes within the selected counties
}
\description{
Get postcodes contained within a local/unitary authority
}
