% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_postcodes.R
\name{get_laua_postcodes}
\alias{get_laua_postcodes}
\title{Get laua postcodes}
\usage{
get_laua_postcodes(postcodes, laua)
}
\arguments{
\item{postcodes}{sf object containing postcodes}

\item{laua}{A character vector of local/unitary authorities}
}
\value{
sf object containing postcodes matching the regex
}
\description{
Get postcodes contained within a local/unitary authority
}
