% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_postcodes.R
\name{get_point_postcodes}
\alias{get_point_postcodes}
\title{Get point postcodes}
\usage{
get_point_postcodes(postcodes, lat, long, crs, distance, units)
}
\arguments{
\item{postcodes}{sf object containing postcodes}

\item{lat}{latitude of point}

\item{long}{longitude of points}

\item{crs}{coordinate reference system}

\item{distance}{maximum distance from point}

\item{units}{units of distance}
}
\value{
sf object containing postcodes within specified distance from point
}
\description{
Get postcodes matching a point definition
}
