% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_postcodes.R
\name{get_polygon_postcodes}
\alias{get_polygon_postcodes}
\title{Get polygon postcodes}
\usage{
get_polygon_postcodes(postcodes, geojson, crs)
}
\arguments{
\item{postcodes}{sf object containing postcodes}

\item{geojson}{A geojson of a single polygon}

\item{crs}{Coordinate reference system of the polygon}
}
\value{
sf object containing postcodes within the polygon
}
\description{
Get postcodes contained with a geojson polygon
}
