% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{get_post_towns}
\alias{get_post_towns}
\title{Get post towns}
\usage{
get_post_towns(postcodes, unique_only = T, sort = T)
}
\arguments{
\item{postcodes}{A character vector of postcodes}

\item{unique_only}{Return only unique post towns}

\item{sort}{Sort post towns}
}
\value{
Post towns
}
\description{
Extract post toens from list of postcodes
}
