% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_postcodes.R
\name{get_postcode_areas}
\alias{get_postcode_areas}
\title{get postcode areas}
\usage{
get_postcode_areas(postcodes)
}
\arguments{
\item{postcodes}{character vector of complete postcodes}

\item{level}{level at which postcode has been}
}
\value{
postcode areas
}
\description{
get postcode areas from character vector of complete postcodes
}
