% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_postcodes.R
\name{get_postcode_districts}
\alias{get_postcode_districts}
\title{get postcode districts}
\usage{
get_postcode_districts(postcodes)
}
\arguments{
\item{postcodes}{character vector of complete postcodes}
}
\value{
postcode districts
}
\description{
get postcode districts from character vector of complete postcodes
}
