% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_postcodes.R
\name{get_postcode_regex}
\alias{get_postcode_regex}
\title{postcode regex}
\usage{
get_postcode_regex(level)
}
\arguments{
\item{level}{level at which postcode has been truncated}
}
\value{
postcode regular expression
}
\description{
Returns complete postcode regular expression
}
