% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_postcodes.R
\name{get_postcode_sectors}
\alias{get_postcode_sectors}
\title{get postcode sectors}
\usage{
get_postcode_sectors(postcodes, separator = " ")
}
\arguments{
\item{postcodes}{character vector of complete postcodes}

\item{seperator}{separator between outward and inward codes}
}
\value{
postcode districts
}
\description{
get postcode sectors from character vector of complete postcodes
}
