% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_postcodes.R
\name{get_postcodes_from_definition}
\alias{get_postcodes_from_definition}
\title{Get postcodes from definitions}
\usage{
get_postcodes_from_definition(postcodes, definition)
}
\arguments{
\item{postcodes}{ONS postcodes}

\item{definiton}{List of area definition}
}
\value{
returns tibble of postcodes
}
\description{
Get list of postcodes from list of definitions
}
