% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_postcodes.R
\name{get_postcodes_from_definition_list}
\alias{get_postcodes_from_definition_list}
\title{Get postcodes from definition list}
\usage{
get_postcodes_from_definition_list(postcodes, definition_list)
}
\arguments{
\item{postcodes}{sf object containing postcode points}

\item{definiton_list}{List of area definitions}
}
\value{
Simple feature collection of postcodes
}
\description{
Function to get postcodes from a list of definitions
}
