% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_postcodes.R
\name{get_regex_postcodes}
\alias{get_regex_postcodes}
\title{Get regex postcodes}
\usage{
get_regex_postcodes(postcodes, regex)
}
\arguments{
\item{postcodes}{ONS postcodes}

\item{definition_list}{A regex definition from yaml file}
}
\value{
sf object containing postcodes matching the regex
}
\description{
Get postcodes matching a regex definition
}
