% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_postcodes.R
\name{get_selected_areas}
\alias{get_selected_areas}
\title{Get selected areas}
\usage{
get_selected_areas(postcodes, selected)
}
\arguments{
\item{postcodes}{ONS postcodes}

\item{selected}{Character vector of postal areas}
}
\value{
sf object containing the selected postcodes
}
\description{
Get directly selected postal areas
}
