% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_postcodes.R
\name{get_selected_districts}
\alias{get_selected_districts}
\title{Get selected districts}
\usage{
get_selected_districts(postcodes, selected)
}
\arguments{
\item{postcodes}{ONS postcodes}

\item{selected}{Character vector of postal districts}
}
\value{
sf object containing the selected postcodes
}
\description{
Get directly selected postal districts
}
