% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_postcodes.R
\name{get_selected_postcodes}
\alias{get_selected_postcodes}
\title{Get selected postcodes}
\usage{
get_selected_postcodes(postcodes, selected)
}
\arguments{
\item{postcodes}{ONS postcodes}

\item{selected}{Character vector of postcodes}
}
\value{
sf object containing the selected postcodes
}
\description{
Get directly selected postcodes
}
