% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_postcodes.R
\name{get_selected_sectors}
\alias{get_selected_sectors}
\title{Get selected sectors}
\usage{
get_selected_sectors(postcodes, selected)
}
\arguments{
\item{postcodes}{ONS postcodes}

\item{selected}{Character vector of postal sectors}
}
\value{
sf object containing the selected postcodes
}
\description{
Get directly selected postal sectors
}
