% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_postcodes.R
\name{get_ward_postcodes}
\alias{get_ward_postcodes}
\title{Get ward postcodes}
\usage{
get_ward_postcodes(postcodes, wards)
}
\arguments{
\item{postcodes}{sf object containing postcodes}

\item{wards}{A character vector of ward codes}
}
\value{
sf object containing postcodes within the selected wards
}
\description{
Get postcodes contained within a local/unitary authority
}
