% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wraper.R
\name{get_wraper}
\alias{get_wraper}
\title{FUNCTION_TITLE}
\usage{
get_wraper(width)
}
\arguments{
\item{width}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
