% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{get_yaml_polygon_postcodes}
\alias{get_yaml_polygon_postcodes}
\title{Get yaml point postcodes}
\usage{
get_yaml_polygon_postcodes(postcodes, yaml_entry)
}
\arguments{
\item{postcodes}{sf object containing postcodes}

\item{yaml_entry}{A point entry from yaml file}
}
\value{
sf object containing postcodes matching the regex
}
\description{
Get postcodes matching a regex entry in yaml file
}
