% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{get_yaml_postcode_list}
\alias{get_yaml_postcode_list}
\title{Get yaml list postcodes}
\usage{
get_yaml_postcode_list(postcodes, area_yaml)
}
\arguments{
\item{postcodes}{ONS postcodes}

\item{area_yaml}{Area defintions from yaml file}
}
\value{
returns tibble of postcodes
}
\description{
Get list of postcodes from yaml file
}
