% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{glaciers}
\alias{glaciers}
\title{Outlines of Icelandic glaciers.}
\format{A data frame with 6469 observations on the following 2 variables.
\describe{ \item{lat}{a numeric vector} \item{lon}{a numeric
vector} }}
\description{
Outlines of Icelandic glaciers.
}
\examples{
data(glaciers)
## maybe str(glaciers) ; plot(glaciers) ...
}
\keyword{datasets}

