% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IATA.R
\name{greatCircle}
\alias{greatCircle}
\title{Great Circle as \link[sp]{SpatialLines} Object}
\usage{
greatCircle(x)
}
\arguments{
\item{x}{\link[sp]{SpatialPoints} object}
}
\value{
Function \code{\link[=greatCircle]{greatCircle()}} returns an \link[sp]{SpatialLines} object.
}
\description{
..
}
