% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{greenland}
\alias{greenland}
\title{Outline of Greenland.}
\format{A data frame with 62416 observations on the following 2 variables.
\describe{ \item{lat}{a numeric vector} \item{lon}{a numeric
vector} }}
\description{
Outline of Greenland.
}
\examples{

data(greenland)
## maybe str(greenland) ; plot(greenland) ...

}
\keyword{datasets}

