% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IATA.R
\name{IATA}
\alias{IATA}
\title{One or More Air Travel Trips using IATA codes}
\usage{
IATA(x)
}
\arguments{
\item{x}{\link[base]{character} scalar or \link[base]{vector}}
}
\value{
Function \code{\link[=IATA]{IATA()}} returns an object of S3 class \code{'IATA'}, which is essentially
a \link[base]{list} of \link[base]{integer} \link[base]{vector}s.
}
\description{
One or more air travel trips using International Air Transport Association (IATA) airport codes.
}
\examples{
IATA('NRT-HNL-YVR')
IATA('CTU-PEK-ICN-JFK, EWR-IAH-LIM')
}
