% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_postcodes.R
\name{identify_postcode_level}
\alias{identify_postcode_level}
\title{identify postcode level}
\usage{
identify_postcode_level(postcodes)
}
\arguments{
\item{postcodes}{Character vector of postcodes}
}
\value{
level of postcodes
}
\description{
Identify if postcodes are area, district, sector or complete
}
