% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{identify_postcode_overlap}
\alias{identify_postcode_overlap}
\title{Identify postcode overlap}
\usage{
identify_postcode_overlap(postcode_list, raise_error = FALSE)
}
\arguments{
\item{postcode_list}{Named list of ONS postcodes for each area}

\item{raise_error}{Raise error if duplicate postcodes found}
}
\description{
Identify postcode overlap in a list of postcodes
}
