\name{init}
\alias{init}

\title{Initiate a geoplot (??)}

\description{
Initiate a geoplot (??).
}

\usage{
init(lat, lon = 0, type = "p", pch = "*", xlim = c(0, 0), ylim = c(0, 0),
  b0 = 65, r = 1.05, country = island, xlab = "Longitude", ylab = "Latitude",
  option = "cut", grid = T, new = F, cont = F, cex = 0.7, col = 1,
  lcont = c(0.13, 0.21), plotit = T, reitur = F, smareitur = F, reittext = F,
  axratio = 1, lwd = 0, axlabels = T, oldpar, projection = "Mercator",
  b1 = 65, dlat = 0, dlon = 0, command = 0, jitter = 0, xaxdist, yaxdist)
}

\arguments{
List of arguments similar to that of \code{geoplot}:
  \item{lat, lon}{Latitude and longitude}
  \item{type}{Plot method}
  \item{pch}{Plot character}
  \item{xlim, ylim}{Plot limits}
  \item{b0}{Base latitude}
  \item{r}{Plot ratio, multiplier on \code{diff(x_or_ylim)}}
  \item{country}{Country to be plotted, default \code{island}}
  \item{xlab, ylab}{Labels for x- and y-axes, default \code{"Longitude",
      "Latitude"}}
  \item{option}{Method of determining plot extent, default \code{"cut"}}
  \item{grid}{Should a grid be drawn, default \code{TRUE}}
  \item{new}{Plot control, default \code{FALSE} adds plot to current plot}
  \item{cont}{For use with contours: should space be reserved for
    contour labels? Default \code{FALSE}}
  \item{cex}{Character size expansion}
  \item{col}{Color, default 1, usually black}
  \item{lcont}{Contour lable space allocation, default c(0.13, 0.21)}
  \item{plotit}{If FALSE plot is only initialized but not plotted. If
    used other programs are used to fill the plot (geolines, geocontour,
    geopolygon etc). Most often used in multiple plots.}
  \item{reitur}{Should the grid be that of statistical rectangles?}
  \item{smareitur}{Should the grid be that of statistical sub--rectangles?}
  \item{reittext}{Should the rectangles be labelled?}
  \item{axratio}{Parameter usually not changed by the user.}
  \item{lwd}{Line width}
  \item{axlabels}{If FALSE no numbers are plotted on the axes. Default
    value is TRUE.}
  \item{oldpar}{The old par--list, from the parent geoplot--call}
  \item{projection}{Projection, default \code{Mercator}}
  \item{b1}{Second latitude for Lambert projection}
  \item{dlat, dlon}{Defines the grid, to make a grid on the axes, 1 is a
    number on axis and a line at every deg. Not usualy set by user.}
  \item{command}{The parent \code{geoplot} command is included, although
    for what purpose isn't quite clear??}
  \item{jitter}{Random jitter to accentuate repeated values. Default no
    jitter (\code{jitter = 0})}
  \item{xaxdist, yaxdist}{Distance from plot to the labels on axes (dist
    or r argument to geoaxis).  Default values \code{0.2, 0.3} but
    higher values mean that axlabels are further away from the plot.
    Further flexibility with axes can be reached by calling geoplot with
    axlabels = FALSE and geoaxis aferwards.}
}

\value{
No value, side effect plotted.
}

\note{
Needs further elaboration, alternatively hide the function.
}

\seealso{
Called by \code{\link{geoplot}} calls \code{\link{Elimcomp}},
\code{\link{findline}}, \code{\link{geoaxis}}, \code{\link{gridaxes}},
\code{\link{gridaxes.Lambert}}, \code{\link{invProj}},
\code{\link{Proj}}.
}

\keyword{hplot}
