\name{inside.reg.bc1}
\alias{inside.reg.bc1}

\title{Determine which bormicon (or gadget) region data belong to.}

\description{
Determine which bormicon (or approximately gadget) region data belong to.
}

\usage{
inside.reg.bc1(data)
}

\arguments{
  \item{data}{Data set with coordinates in components \code{lat, lon}.}
}

\value{
Returns original data with area code as an added component \code{area}.
}

\note{
Needs further elaboration ?
}

\seealso{
Calls \code{\link{geoinside}}, called by \code{\link{inside.reg.bc}} and \code{\link{Reitur2Svaedi1to10}}. Data set \code{\link{reg.bc}} with bormicon-area outlines is used.
}

\keyword{manip}
