\name{invlambert}
\alias{invlambert}

\title{Inverse Lambert projection}

\description{
Inverse Lambert projection.
}

\usage{
invlambert(x, y, lat0, lon0, lat1, scale = "km", old = F)
}

\arguments{
  \item{x, y}{The input data to be inversely projected as two vectors}
  \item{lat0}{First latitude defining of the projection}
  \item{lon0}{Longitude defining the projection}
  \item{lat1}{Second latitude defining the projection}
  \item{scale}{Scale of the input data, "km" default, redundant ??}
  \item{old}{Old method, seldom used???}
}

\value{
List of components:
  \item{lat, lon }{Coordinates in latitude and longitude}
  \item{x, y}{Input coordinates (projected)}
  \item{scale}{Scale}
  \item{projection}{Projection (stated redunantly)}
  \item{lat0, lon0, lat1}{Defining lats and lon}
is returned invisibly.
}

\note{
Needs elaboration and perhaps documenting with lambert in the same doc-file.
}

\seealso{
Called by \code{\link{invProj}}.
}

\keyword{manip}
