% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{island}
\alias{island}
\title{Icelandic coastline.}
\format{A data frame with 1323 observations on the following 2 variables.
\describe{ \item{lat}{a numeric vector} \item{lon}{a numeric
vector} }}
\source{
Unknown.
}
\description{
Icelandic coastline from unknown source.
}
\keyword{datasets}
