\name{labels1}
\alias{labels1}

\title{Label plots}

\description{
Label plots with categories.
}

\usage{
labels1(cont, digits, colors, xlim = c(0, 1), ylim = c(0, 1), fill = F, 
  minsym = "<", label.resolution = 0, labtxt = NULL, first.color.trans = T, 
  mai = c(0, 1, 0, 1), leftrat = 0.1)
}

\arguments{
  \item{cont}{Contour ?}
  \item{digits}{Number of digits to use in labels}
  \item{colors}{Colors ?}
  \item{xlim, ylim}{Limits ?}
  \item{fill}{Fill with colors?}
  \item{minsym}{Minimum symbol (for lowest category?)?}
  \item{label.resolution}{Label resolution ?}
  \item{labtxt}{Label text ?}
  \item{first.color.trans}{Should first color be transparent? Default TRUE}
  \item{mai}{Margins in inches?}
  \item{leftrat}{Left ratio (giving space for labels??)??}
}

\value{
No value, labels added to current plot.
}

\note{
Needs elaboration, merge documentation with \code{labels2}, and possibly others?
}

\seealso{
Called by \code{\link{colsymbol}}, \code{\link{geocontour.fill}} and \code{\link{reitaplott}}.
}

\keyword{aplot}
