\name{labels_line}
\alias{labels_line}

\title{Labels line ?}

\description{
Labels line ?.
}

\usage{
labels_line(cont, digits, colors, lty, lwd, xlim = c(0, 1), ylim = c(0, 1), linew = F)
}

\arguments{
  \item{cont}{Contours?}
  \item{digits}{Number of digits}
  \item{colors}{Colors}
  \item{lty}{Line types}
  \item{lwd}{Line widths}
  \item{xlim, ylim}{Limit}
  \item{linew}{Linewidth for some reason not lwd?}
}

\value{
No value, labels added to current plot.
}

\note{
Needs elaboration, possibly merge documentation with others label functions?
}

\seealso{
Called by \code{\link{geocontour}}.
}

\keyword{aplot}
