\name{labels_size}
\alias{labels_size}

\title{Label symbols of given size}

\description{
Labels of given size.
}

\usage{
labels_size(cont, digits, sizes, xlim = c(0, 1), ylim = c(0, 1), fill = F, n, rat, minsym = "<", label.resolution = 0, open = F, lwd = 1, col = 1)
}

\arguments{
  \item{cont}{Contours}
  \item{digits}{Number of digits to use in labels}
  \item{sizes}{Sizes (of what?)?}
  \item{xlim, ylim}{Limits}
  \item{fill}{Fill? Default FALSE}
  \item{n}{Number of ??}
  \item{rat}{Ratio of ??}
  \item{minsym}{Minimum symbol for lowest category in labels}
  \item{label.resolution}{Label resolution ?}
  \item{open}{Open legend/label, default FALSE}
  \item{lwd}{Line width}
  \item{col}{Color}
}

\value{
No value, labels added to current plot.
}

\note{
Needs further elaboration, document with other labelling functions??
}

\seealso{
Called by \code{\link{colsymbol}}.
}

\keyword{aplot}
