% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{leaflet_mrk}
\alias{leaflet_mrk}
\title{\link[leaflet]{leaflet} with Markers}
\usage{
leaflet_mrk(coords, popup = rownames(coords), ...)
}
\arguments{
\item{coords}{2-column \link[base]{matrix} of markers, \emph{longitude} on the 1st column and \emph{latitude} on the 2nd column}

\item{popup}{\link[base]{character} \link[base]{vector}, the popup text.  Default value is the \link[base]{rownames} of \code{coords}}

\item{...}{additional parameters, currently not in use}
}
\value{
The function \code{\link[=leaflet_mrk]{leaflet_mrk()}} returns a \link[leaflet]{leaflet} object.
}
\description{
To create a \link[leaflet]{leaflet} with markers from coordinates.
}
