% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{leaflet_plg}
\alias{leaflet_plg}
\title{\link[leaflet]{leaflet} with Polygons}
\usage{
leaflet_plg(
  sf,
  fillColor = "royalblue",
  fillOpacity = 0.1,
  weight = 2,
  color = "white",
  opacity = 1,
  ...
)
}
\arguments{
\item{sf}{an \code{sf} object}

\item{fillColor, fillOpacity, weight, color, opacity, ...}{additional parameters of the function \link[leaflet]{addPolygons}}
}
\description{
To create a \link[leaflet]{leaflet} with polygons.
}
