% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{leaflet_popup}
\alias{leaflet_popup}
\title{\link[leaflet]{leaflet} with Popup}
\usage{
leaflet_popup(coords, ...)
}
\arguments{
\item{coords}{2-column \link[base]{matrix} of popup
\emph{longitude} (1st column) and \emph{latitude} (2nd column)}

\item{...}{additional parameters, currently not in use}
}
\description{
..
}
