% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litir.R
\name{litir}
\alias{litir}
\title{Display palette in effect}
\usage{
litir(n)
}
\arguments{
\item{n}{Number of columns and rows of colors to display}
}
\value{
No value returned, plots an \code{n} by \code{n} grid of colors.
}
\description{
A grid of colors is plotted.
}
\details{
Palette in effect gets repeated for \code{n} greater than
\code{length(palette())}.
}
\note{
Simpler version of \code{\link{Rlitir}}, one of those should do,
possibly with a name change.
}
\seealso{
\code{\link{Rlitir}}
}
\keyword{colors}

