% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_n_vanilla.R
\name{lower_n}
\alias{lower_n}
\title{The \eqn{n}th Lower Edge(s) of a Matrix}
\usage{
lower_n(x, n = seq_len(.dim[1L] - 1L))
}
\arguments{
\item{x}{\link[base]{matrix}}

\item{n}{\link[base]{integer} scalar or \link[base]{vector},
order(s) of the lower edge(s)}
}
\value{
Function \code{\link[=lower_n]{lower_n()}} returns a \link[base]{logical} \link[base]{matrix}.
}
\description{
..
}
\details{
Function \code{\link[=lower_n]{lower_n()}} extends \link[base]{lower.tri}, so that
the \link[base]{logical} indices of \eqn{n}th lower-edge(s) elements are returned.
}
\examples{
# ?euro.cross
dim(euro.cross) # square matrix
lower_n(euro.cross) # I love dimnames 
lower.tri(euro.cross) # no dimnames
lower_n(euro.cross, 1:2)

dim(VADeaths)
lower_n(VADeaths, n = 1:2)

(x = array(1, dim = c(1, 1)))
lower_n(x) # exception

}
