% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loyalty.R
\docType{class}
\name{loyalty}
\alias{loyalty}
\alias{loyalty-class}
\title{\linkS4class{loyalty}}
\description{
\linkS4class{loyalty}
}
\section{Slots}{

\describe{
\item{\code{program}}{\link[base]{character} scalar, airline program to which the flight is credited (\strong{not} carrier of the flight)}

\item{\code{goal}}{\link[base]{numeric} \link[base]{vector}}

\item{\code{reward}}{\link[base]{numeric} \link[base]{vector}}

\item{\code{status}}{\link[base]{numeric} \link[base]{vector}}
}}

\examples{
(JSM24 = new(Class = 'airfare', carrier = 'AA', depart = 'PHL', arrive = 'PDX',
  code = 'J', basefare = 276.28, tax = 35.82, upgrade = 241.88))
earnAA(JSM24)
earnAS(JSM24)
earnBA(JSM24)
}
