% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_dist2gc.R
\name{moving_dist2gc}
\alias{moving_dist2gc}
\alias{moving_dist2gc.iata}
\alias{moving_dist2gc.iatalist}
\alias{moving_dist2gc.sf}
\alias{moving_dist2gc.matrix}
\title{Moving Cross Track Distance}
\usage{
moving_dist2gc(x, ...)

\method{moving_dist2gc}{iata}(x, ...)

\method{moving_dist2gc}{iatalist}(x, ...)

\method{moving_dist2gc}{sf}(x, ...)

\method{moving_dist2gc}{matrix}(x, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{additional parameters of the function \link[geosphere]{dist2gc}}
}
\description{
Moving Cross Track Distance
}
