% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gdcVolcanoPlot.R
\name{my_gdcVolcanoPlot}
\alias{my_gdcVolcanoPlot}
\title{FUNCTION_TITLE}
\usage{
my_gdcVolcanoPlot(deg.all, fc = 2, pval = 0.05)
}
\arguments{
\item{deg.all}{PARAM_DESCRIPTION}

\item{fc}{PARAM_DESCRIPTION, Default: 2}

\item{pval}{PARAM_DESCRIPTION, Default: 0.01}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{lims}},\code{\link[ggplot2]{geom_abline}},\code{\link[ggplot2]{ggtheme}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}}
}
