% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{nonsetpar}
\alias{nonsetpar}
\title{Some parameter settings.}
\format{The format is: chr [1:8] "omd" "oma" "omi" "fin" "fig" "mfcol"
"mfrow" ...}
\description{
Needs elaboration or to be removed.
}
\examples{

data(nonsetpar)
## maybe str(nonsetpar) ; plot(nonsetpar) ...

}
\keyword{datasets}
