% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_postcodes.R
\name{order_postcodes}
\alias{order_postcodes}
\title{order postcodes}
\usage{
order_postcodes(postcodes, level = NULL)
}
\arguments{
\item{postcodes}{character vector of postcodes}

\item{level}{level at which postcode has been truncated}
}
\value{
integer vector of correct order
}
\description{
order postcodes by area, district, sector and unit
}
