\name{paint.window}
\alias{paint.window}

\title{Paint window for label?}

\description{
Paint window for label based on lat and lon?.
}

\usage{
paint.window(listi, col = 0, border = T, poly = T, col.names = c("lon", "lat"))
}

\arguments{
  \item{listi}{Label location ?}
  \item{col}{Color, not used ?}
  \item{border}{Should border be drawn?}
  \item{poly}{Should label be opaque ?}
  \item{col.names}{Column names containing label coordinates ?}
}

\value{
No value, lines and/or polygon added to current geoplot.
}

\note{
Needs elaboration and possibly merging with paint.window doc-file. Argument \code{col} has no effect.
}

\seealso{
Called by \code{\link{colsymbol}}, \code{\link{geocontour}}, \code{\link{geocontour.fill}}, \code{\link{geosymbols}} and \code{\link{reitaplott}}; calls \code{\link{Proj}}.
}

\keyword{aplot}
