\name{paint.window.x}
\alias{paint.window.x}

\title{Paint window for label ?}

\description{
Paint window for label based on projected coordinates?.
}

\usage{
paint.window.x(listi, col = 0, border = T, poly = T)
}

\arguments{
  \item{listi}{Label location ?}
  \item{col}{Color, not used ?}
  \item{border}{Should border be drawn?}
  \item{poly}{Should label be opaque ?}
}

\value{
No value, lines and/or polygon for label added to current geoplot.
}

\note{
Needs elaboration and possibly merging with paint.window doc-file. Argument \code{col} has no effect.
}

\seealso{
Called by \code{\link{geocontour}}.
}

\keyword{aplot}
