\name{pdistx}
\alias{pdistx}

\title{Euclidian distance}

\description{
Euclidian distance.
}

\usage{
pdistx(y, x, y1, x1)
}

\arguments{
  \item{y, x}{Coordinate vectors of starting positions}
  \item{y1, x1}{Coordinate vectors of end positions}
}

\value{
Euclidian distance
}

\note{
Check order of coordinates (?), document distance functions together?
}

\seealso{
Called by \code{\link{variogram}}.
}

\keyword{manip}
