% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_postcodes.R
\name{perform_postcode_operations}
\alias{perform_postcode_operations}
\title{perform postcode operations}
\usage{
perform_postcode_operations(postcode_list, operations)
}
\description{
Function to perform all operations on postcodes
}
