% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.iata.R
\name{plot.iata}
\alias{plot.iata}
\title{Draw \code{'iata'} using \CRANpkg{plotly}}
\usage{
\method{plot}{iata}(
  x,
  ...,
  map = plot_geo(),
  col = pal_hue()(n = 1L),
  geo = list(resolution = 50, showland = TRUE, landcolor = toRGB("grey97"), showlakes =
    TRUE, lakecolor = toRGB("lightblue"), showrivers = TRUE, rivercolor =
    toRGB("lightblue"), riverwidth = 0.5, showcountries = TRUE, countrycolor =
    toRGB("grey50"), countrywidth = 0.5, showocean = TRUE, oceancolor = toRGB("white"),
    coastlinecolor = toRGB("grey50"), coastlinewidth = 0.5, lonaxis = list(showgrid =
    TRUE, gridcolor = toRGB("gray80"), gridwidth = 0.5), lataxis = list(showgrid = TRUE,
    gridcolor = toRGB("gray80"), 
     gridwidth = 0.5), projection = list(type =
    "orthographic", rotation = list(lon = -100, lat = 40)))
)
}
\arguments{
\item{x}{an \code{'iata'}}

\item{...}{..}

\item{map}{an htmlwidget of world map, default is the return of function
\link[plotly]{plot_geo}}

\item{col}{\link[base]{character} scalar}

\item{geo}{a \link[base]{list}, see \url{https://plotly.com/r/reference/layout/geo/} for detail}
}
\description{
Draw \code{'iata'} using \CRANpkg{plotly}
}
\keyword{internal}
