% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.iata.R
\name{plot.iatalist}
\alias{plot.iatalist}
\title{Draw \code{'iatalist'} using \CRANpkg{plotly}}
\usage{
\method{plot}{iatalist}(x, ..., map = plot_geo())
}
\arguments{
\item{x}{\code{'iatalist'}}

\item{...}{..}

\item{map}{..}
}
\description{
..
}
\references{
\url{https://plotly.com/r/lines-on-maps/}
}
\keyword{internal}
