% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrich.R
\name{plot_enrich}
\alias{plot_enrich}
\title{plot enrichment analysis}
\usage{
plot_enrich(enrich_df, title = "Statistics of Enrichment")
}
\arguments{
\item{enrich_df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{arrange}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{rename}},\code{\link[dplyr]{select}}
 \code{\link[janitor]{clean_names}}
 \code{\link[utils]{head}}
 \code{\link[tidyr]{separate}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{scale_colour_gradient}},\code{\link[ggplot2]{ggtheme}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{scale_x_discrete}},\code{\link[ggplot2]{theme}}
 \code{\link[forcats]{fct_reorder}}
}
