\name{plot_nogrid}
\alias{plot_nogrid}

\title{Grid control ?}

\description{
Grid or border control of some sort?
}

\usage{
plot_nogrid(o, xgr, ygr, col)
}

\arguments{
  \item{o}{o, some outer boundary?}
  \item{xgr}{xgr ?}
  \item{ygr}{ygr ?}
  \item{col}{Color}
}

\value{
No value, some lines added to current geoplot.
}

\note{
Needs elaboration.
}

\seealso{
Called by \code{\link{gridaxes}}.
}

\keyword{aplot}
