% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_postcodes}
\alias{plot_postcodes}
\title{Plot postcodes}
\usage{
plot_postcodes(tb, center_zoom = NULL)
}
\arguments{
\item{tb}{Table containing plot data}
}
\value{
Return plot
}
\description{
Calculate cnter and zoom for a mapbox plot
}
