% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{postcode_plot_data_from_list}
\alias{postcode_plot_data_from_list}
\title{Postcode plot data from list}
\usage{
postcode_plot_data_from_list(
  postcodes,
  postcode_col,
  postcode_list,
  warn_missing = TRUE
)
}
\arguments{
\item{postcodes}{Table of current postcodes}

\item{postcode_col}{Column from postcode table to use for matching}

\item{postcode_list}{List of postcodes}

\item{warn_missing}{Warn if missing postcodes are supplied}
}
\value{
A table for plotting postcodes
}
\description{
Generate postcode plot data from a list of postcodes
}
