% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{postcol}
\alias{postcol}
\title{Old palette.}
\format{The format is: num [1:169, 1:3] 0.667 0.667 0.660 0.654 0.647 ...  -
attr(*, "dimnames")=List of 2 ..$ : chr [1:169] "1" "2" "3" "4" ...  ..$ :
chr [1:3] "H" "S" "V"}
\description{
Soon to be deprecated, still referred to by \code{geopalette}
and \code{geocontour.fill}..
}
\examples{
data(postcol)
## maybe str(postcol) ; plot(postcol) ...
}
\keyword{datasets}

