\name{prepare.line}
\alias{prepare.line}

\title{Prepare a line ?}

\description{
Prepare a line for some action ?
}

\usage{
prepare.line(x)
}

\arguments{
  \item{x}{Line vector ?}
}

\value{
Returns a list with components:
  \item{lx1}{lx1 ?}
  \item{lx2}{lx2 ?}
  \item{nlx}{nlx ?}
}

\note{
Needs elaboration.
}

\seealso{
Called by \code{\link{cut_multipoly}} and \code{\link{findline}}.
}

\keyword{manip}
