% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.iata.R
\name{print.iatalist}
\alias{print.iatalist}
\title{print.iatalist}
\usage{
\method{print}{iatalist}(x, ...)
}
\arguments{
\item{x}{\code{'iatalist'}}

\item{...}{..}
}
\description{
print.iatalist
}
\keyword{internal}
