% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{read_geojson_polygon}
\alias{read_geojson_polygon}
\title{Read geojson polygon}
\usage{
read_geojson_polygon(geojson, input_crs)
}
\arguments{
\item{geojson}{geojson string or path to geojson file}

\item{input_crs}{input coordinate reference system}
}
\value{
simple feature collection containing polygons
}
\description{
Read geojson polygon and convert to 4326 CRS
}
