% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_postcodes.R
\name{read_ons_gb_postcodes}
\alias{read_ons_gb_postcodes}
\title{Read ons gb postcodes}
\usage{
read_ons_gb_postcodes(path, max_osgrdind, min_doterm, crs)
}
\arguments{
\item{path}{Path to ONSPD file}

\item{max_osgrdind}{Maximum allowable ONS grid reference quality indicator}

\item{min_doterm}{Minimum date of termination allowed}
}
\value{
simple feature collection containing polygons
}
\description{
Read GB postcodes in ONS postcode directory (ONSPD) file
}
