% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_postcodes.R
\name{read_ons_postcodes}
\alias{read_ons_postcodes}
\title{Read ons postcodes}
\usage{
read_ons_postcodes(path, rm_terminated, rm_nogrid)
}
\arguments{
\item{path}{Paths to postcode CSV}
}
\description{
Read ons postcode csv, filter and convert to sf object
}
