\name{rectangleArea}
\alias{rectangleArea}
\alias{drArea}
\alias{mrArea}
\alias{hrArea}
\alias{srArea}
\alias{rArea}

\title{
Given rectangle code return area in square kilometers or nautical miles
}
\description{
Rectangle area is returned through a call to
\link{\code{rectangelPerimeter}}-fuyntions and \link[geo]{\code{geoarea}}.
}
\usage{
rPeri(r)
srPeri(sr)
}

\arguments{
  \item{r, sr}{rectangle code}

}

\keyword{manip}
\keyword{arith}
