% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SegmentWithArrow.R
\name{SegmentWithArrow}
\alias{SegmentWithArrow}
\title{Plot line segment with arrow at the end.}
\usage{
SegmentWithArrow(pos, angle = 15, size = 0.2, col = "blue", lwd = 2)
}
\arguments{
\item{pos}{Segment positions.}

\item{angle}{(Half-)Angle (< 90) determining arrow sharpness.}

\item{size}{Size of arrow.}

\item{col}{Color of arrow.}

\item{lwd}{Line width of segment.}
}
\description{
Plot line segment with arrow at the end.
}
\note{
Needs further checking and elaboration.
}
\keyword{aplot}

