\name{selectedpar}
\alias{selectedpar}

\title{Parameters manipulation ?}

\description{
Parameters manipulation ?
}

\usage{
selectedpar()
}

\value{
Returns the names of some selected parameters.
}

\note{
Needs elaboration, 
}

\seealso{
Called by the majority of the graphical functions in geo, calls \code{\link{Elimcomp}}.
}

\keyword{device}
