\name{shadeborder}
\alias{shadeborder}

\title{Shade border ?}

\description{
Shade border ?
}

\usage{
shadeborder(reg, lat, lon, col = 0, col.names = c("lon", "lat"))
}

\arguments{
  \item{reg}{Region}
  \item{lat}{Latitude ?}
  \item{lon}{Longitude ?}
  \item{col}{Color, not used ?}
  \item{col.names}{Column names containing coordinates, default \code{lat} and \code{lon}.}
}

\value{
No value, addes shadedborder (with call to \code{\link{lines}}) to contoured geoplot.
}

\note{
Needs elaboration. Color argument not used, but fixed values given to 2 calls to \code{lines}.
}

\seealso{
Called by \code{\link{geocontour}}, calls \code{\link{Proj}}.
}

\keyword{aplot}
