\name{shading1}
\alias{shading1}

\title{Shading of geoplots?}

\description{
Shading of geoplots?
}

\usage{
shading1(cont, digits, colors, xlim = c(0, 1), ylim = c(0, 1), 
  fill = F, angle, rotate, cex, rat, minsym = "<")
}

\arguments{
  \item{cont}{Contours?}
  \item{digits}{Number of digits?}
  \item{colors}{Colors?}
  \item{xlim, ylim}{Limits?}
  \item{fill}{Fill?}
  \item{angle}{Angle?}
  \item{rotate}{Rotate?}
  \item{cex}{Character expansion?}
  \item{rat}{Ratio?}
  \item{minsym}{Minimum symbol on label?}
}

\value{
No value, shades current geoplot (label?) in some way?
}

\note{
Needs elaboration.
}

\seealso{
Called by \code{\link{colsymbol}} and \code{\link{reitaplott}}.
}

\keyword{aplot}
