\name{skipta.texta}
\alias{skipta.texta}

\title{Translate characters in column names}

\description{
Translate characters in column names
}

\usage{
skipta.texta(txt, char = "_", replchar = ".")
}

\arguments{
  \item{txt}{Character vector of inputs}
  \item{char}{Character to be replaced}
  \item{replchar}{Character to replace with}
}

\value{
A character vector of the same length as \code{txt}.
}

\note{
Could be deprecated, use \code{\link{chartr}} instead.
}

\seealso{
Called by \code{\link{pre2s}}.
}

\keyword{manip}
