\name{SMB.std.background}
\alias{SMB.std.background}

\title{Standard background for the SMB.}

\description{
Plots standard background for the SMB survey, aka Icelandic Groundfish Survey.
}

\usage{
SMB.std.background(depth, depthcol = 1, depthlty = 1, depthlwd = 1, 
  eyjar, depthlab, depthlabcsi = 0.12, ...)
}

\arguments{
  \item{depth}{Vector of isobaths to draw, default none}
  \item{depthcol}{Colors for isobaths}
  \item{depthlty}{Line types for isobaths}
  \item{depthlwd}{Line widths for isobaths}
  \item{eyjar}{Additional data to plot, e.g. islands}
  \item{depthlab}{Label depths? Defaults to FALSE}
  \item{depthlabcsi}{Depthlabel size}
  \item{\dots}{Additional arguemnts to \code{geoplot} and \code{geolines}}
}

\value{
No value returned, plots standard layout SMB/IGFS plot.
}

\note{
Depth labelling etc might be cleaned up, use of \code{eyjar} argument could change, calls data object \code{depthloc} with a few positions for isobath labels.
}

\seealso{
\code{\link{geoplot}}, \code{\link{gbplot}}, \code{\link{geolines}}, \code{\link{eyjar}}
}

\keyword{hplot}
