% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_postcodes.R
\name{sort_postcodes}
\alias{sort_postcodes}
\title{sort postcodes}
\usage{
sort_postcodes(postcodes, level = NULL)
}
\arguments{
\item{postcodes}{character vector of postcodes}

\item{level}{level at which postcode has been truncated}
}
\value{
sorted postcodes
}
\description{
sort postcodes by area, district, sector and unit
}
