% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherical.R
\name{spherical}
\alias{spherical}
\title{Spherical??}
\usage{
spherical(rang1, sill, nugget, x)
}
\arguments{
\item{rang1, sill, nugget}{Parameters of a variogram.}

\item{x}{The data.}
}
\value{
Fitted values?
}
\description{
'Spherical' of a variogram?
}
\note{
Needs elaboration, if this is a necessary function.
}
\seealso{
Some variogram stuff.
}
\keyword{arith}
