% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postal_sectors.R
\name{split_postcodes_by_sector}
\alias{split_postcodes_by_sector}
\title{Split postcodes by sector}
\usage{
split_postcodes_by_sector(postcodes)
}
\arguments{
\item{postcodes}{Vector of postcodes}
}
\description{
Split postcodes by their postal sector
}
