% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_styleURL.R
\name{styleURL}
\alias{styleURL}
\title{Style URL}
\usage{
styleURL(url_, text_)
}
\arguments{
\item{url_}{\link[base]{character} scalar or \link[base]{vector}}

\item{text_}{(optional) \link[base]{character} scalar or \link[base]{vector}}
}
\description{
Style URL
}
\examples{
cat(styleURL(c('www.google.com', 'www.apple.com')), sep = '\n')

}
\references{
\url{https://github.com/rstudio/rstudio/issues/1941}
}
