\name{subplot}
\alias{subplot}

\title{Add subplot}

\description{
Add subplot.
}

\usage{
subplot(fun, x, y, size = c(1, 1), vadj = 0.5, hadj = 0.5, pars)
}

\arguments{
  \item{fun}{Graphcical function to call}
  \item{x, y}{Coordinates}
  \item{size}{Subplot size (inches?)}
  \item{vadj}{Vertical adjustment}
  \item{hadj}{Horizontal adjustment}
  \item{pars}{Parameters to set?}
}

\value{
Returns parameter list invisibly, side effect subplot added to current plot.
}

\note{
Needs elaboration.
}

\seealso{
Called by \code{\link{geosubplot}}, calls \code{\link{selectedpar}}.
}

\keyword{aplot}
