% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_postcodes.R
\name{truncate_postcodes}
\alias{truncate_postcodes}
\title{truncate postcodes}
\usage{
truncate_postcodes(postcodes, level, unique, sort)
}
\arguments{
\item{postcodes}{character vector of postcodes}

\item{level}{one of area, district or sector}
}
\value{
truncated postcodes
}
\description{
truncate postcodes to area, district or sector
}
