% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{turn_IATA}
\alias{turn_IATA}
\title{Draw \code{'IATA'} using \CRANpkg{plotly}}
\usage{
turn_IATA(object, ...)
}
\arguments{
\item{object}{\code{'IATA'}}

\item{...}{..}
}
\description{
..
}
\references{
\url{https://plotly.com/r/lines-on-maps/}
}
