% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{twohmiles}
\alias{twohmiles}
\title{The Icelandic EEZ.}
\format{A data frame with 88 observations on the following 2 variables.
\describe{ \item{lat}{a numeric vector} \item{lon}{a numeric
vector} }}
\source{
Some official source hopefully.
}
\usage{
twohmiles
}
\description{
Outline of the Icelandic EEZ established somehow.
}
\note{
Needs checking, e.g. has recent agreement with Faroes been
implemented. Compare with what's on \url{www.marineregions.org}.
}
\references{
Needed.
}
\keyword{datasets}

