\name{adapt}
\alias{adapt}

\title{Adapts geographical positions}

\description{
Adapts geographical positions for further geo-use.
}
\usage{
adapt(reg.lat, reg.lon, projection = "Mercator")
}

\arguments{
  \item{reg.lat}{Latitude or y-coordinate}
  \item{reg.lon}{Longitude or x-coordinate}
  \item{projection}{Projection, default "Mercator", "none" denotes x/y coordinates}
}

\value{
Returns a list of either:
  \item{x, y}{x- and y-positions}
or:
  \item{lat, lon}{Latitude and longitude}
and:
  \item{lxv}{index of uninterupted/contiguous positions}
}

\note{
Needs further elaboration, this function is called by \code{geoarea.old}, \code{geoinside}, \code{inside} and \code{pointkriging}.
}

\keyword{manip}
