% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{add_rest_of_gb}
\alias{add_rest_of_gb}
\title{add rest of gb}
\usage{
add_rest_of_gb(postcodes, postcode_list)
}
\arguments{
\item{postcodes}{ONS postcodes}

\item{postcode_list}{Named list of ONS postcodes for each area}
}
\value{
updated postcode list
}
\description{
add rest of gb to postcode list
}
