% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.grd.R
\name{adjust.grd}
\alias{adjust.grd}
\title{Adjusts grid for plot in Lambert projection.}
\usage{
adjust.grd(ply, rat = 0.025)
}
\arguments{
\item{ply}{gridline??}

\item{rat}{curvature or slant???}
}
\value{
Returns diagonal or curved lines for use in grid on Lambert
projected plots.
}
\description{
Adjusts grid for plot in Lambert projection.
}
\note{
Needs elaboration. Utility, internal, might be hidden?
}
\seealso{
\code{\link{findline}}, \code{\link{gridaxes.Lambert}}
}
\keyword{aplot}
\keyword{manip}

