\name{adjust.grd}
\alias{adjust.grd}

\title{Adjusts grid for plot in Lambert projection.}

\description{
Adjusts grid for plot in Lambert projection.
}

\usage{
adjust.grd(ply, rat = 0.025)
}

\arguments{
  \item{ply}{gridline??}
  \item{rat}{curvature or slant???}
}

\value{
Returns diagonal or curved lines for use in grid on Lambert projected plots.
}

\note{
Needs elaboration. Utility, internal, might be hidden?
}

\seealso{
\code{\link{findline}}, \code{\link{gridaxes.Lambert}}
}

\keyword{manip}
\keyword{aplot}
