% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ana099.R
\name{ana.099.im}
\alias{ana.099.im}
\title{ANA Round Trip from US}
\usage{
ana.099.im(data, US, min. = 21L, max. = 45L, ...)
}
\arguments{
\item{data}{\link[base]{data.frame}, downloaded from \url{https://ana.099.im}}

\item{US}{\link[base]{character} scalar or \link[base]{vector},
US airport for departure/arrival}

\item{min., max.}{\link[base]{integer} scalar, minimum and maximum days
between two departure dates.}

\item{...}{..}
}
\description{
ANA Round Trip from US
}
\details{
Must download latest \code{.csv} file from \url{https://ana.099.im},
upper-right corner of table, download symbol.
}
\examples{
\dontrun{
# Must download latest `.csv` file from https://ana.099.im
tm = Sys.time(); attr(tm, 'tzone') = 'GMT'
(fl = list.files(path = '~/Downloads', pattern = paste0('^', as.Date(tm), '.*_export.csv'), 
  full.names = TRUE))
dim(ana <- read.csv(file = sort(fl, decreasing = TRUE)[1L], header = TRUE))
c(ana$departure, ana$arrival) |> 
 setdiff('TYO') |>
 paste(collapse = '-') |>
 IATA() |>
 as.leaflet()
ana.099.im(ana, US = c('JFK', 'IAD'), min. = 20, max. = Inf)
ana.099.im(ana, US = c('SFO', 'LAX'), min. = 20, max. = Inf)
ana.099.im(ana, US = c('SEA', 'YVR')) #, min. = 20, max. = Inf)
ana.099.im(ana, US = c('IAH'))
ana.099.im(ana, US = c('MEX'), min. = 20L, max. = Inf)
}
}
