% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{as.leaflet}
\alias{as.leaflet}
\alias{as.leaflet.SpatialPoints}
\alias{as.leaflet.sf}
\alias{as.leaflet.iatalist}
\title{Convert to \CRANpkg{leaflet}}
\usage{
as.leaflet(x, ...)

\method{as.leaflet}{SpatialPoints}(x, ...)

\method{as.leaflet}{sf}(x, ...)

\method{as.leaflet}{iatalist}(x, popup = rownames(ap), ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{popup, ...}{additional parameters of the function \code{\link[=leaflet_mrk]{leaflet_mrk()}}}
}
\value{
The function \code{\link[=as.leaflet]{as.leaflet()}} returns a \link[leaflet]{leaflet} object.
}
\description{
To convert an R object into a \link[leaflet]{leaflet}.
}
\note{
1 degree approx 69 miles.

Consider geocoding via \CRANpkg{tidygeocoder}.
}
\examples{
'EWR-PHL-JFK-IAD' |> as.iata() |> as.leaflet()
'EWR-PHL-JFK-IAD, DFW-IAH' |> as.iata() |> as.leaflet()
}
\references{
\url{http://rstudio.github.io/leaflet/}
\url{https://developers.google.com/maps/documentation/geocoding/}
\url{https://shiny.posit.co/r/gallery/}
}
