% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{as.leaflet}
\alias{as.leaflet}
\alias{as.leaflet.sf}
\alias{as.leaflet.iatalist}
\title{Convert to \CRANpkg{leaflet}}
\usage{
as.leaflet(x, ...)

\method{as.leaflet}{sf}(x, ...)

\method{as.leaflet}{iatalist}(x, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{additional parameters of the functions \link[leaflet]{addMarkers} and \link[leaflet]{addPolygons}.}
}
\value{
The function \code{\link[=as.leaflet]{as.leaflet()}} returns a \link[leaflet]{leaflet} object.
}
\description{
To convert an R object into a \link[leaflet]{leaflet}.
}
\note{
1 degree approx 69 miles.
}
\examples{
'EWR-PHL-JFK-IAD' |> as.iata() |> as.leaflet(popup = ~ shortnm)
'EWR-PHL-JFK-IAD, DFW-IAH' |> as.iata() |> as.leaflet(popup = ~ shortnm)
}
\references{
\url{http://rstudio.github.io/leaflet/}
\url{https://shiny.posit.co/r/gallery/}
}
