% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{as.leaflet}
\alias{as.leaflet}
\alias{as.leaflet.iatalist}
\alias{as.leaflet.SpatialPoints}
\title{Convert to \CRANpkg{leaflet}}
\usage{
as.leaflet(x, ...)

\method{as.leaflet}{iatalist}(x, ...)

\method{as.leaflet}{SpatialPoints}(x, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{potential parameters, currently not in use}
}
\value{
The function \code{\link[=as.leaflet]{as.leaflet()}} returns a \link[leaflet]{leaflet} object.
}
\description{
Convert to \CRANpkg{leaflet}
}
\details{
Google geocode API has a query limit, thus 'character' location name such as 'white house'
is not allowed.
}
\note{
1 degree approx 69 miles
}
\examples{
'EWR-PHL-JFK-IAD' |> as.iata() |> as.leaflet()
}
\references{
\url{http://rstudio.github.io/leaflet/}
\url{https://developers.google.com/maps/documentation/geocoding/}
\url{http://shiny.rstudio.com/gallery}
}
