% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{as.leaflet}
\alias{as.leaflet}
\title{Print using \CRANpkg{leaflet}}
\usage{
as.leaflet(x, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{potential parameters, currently not in use}
}
\value{
Function \link{as.leaflet} returns a \link[leaflet]{leaflet} object.
}
\description{
1 degree approx 69 miles
}
\details{
Google geocode API has a query limit, thus 'character' location name such as 'white house'
is not allowed.
}
\examples{
as.leaflet(IATA('EWR-PHL-JFK-IAD'))

}
\references{
\url{http://rstudio.github.io/leaflet/}
\url{https://developers.google.com/maps/documentation/geocoding/}
\url{http://shiny.rstudio.com/gallery/tags/leaflet/}
}
