% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elimcomp.R
\name{Elimcomp}
\alias{Elimcomp}
\title{Parameters manipulation ?}
\usage{
Elimcomp(parlist)
}
\arguments{
\item{parlist}{List of parameters ?}
}
\value{
Manipulated (shortened ?) parameterlist
}
\description{
Parameters manipulation?
}
\note{
Needs elaboration, \code{selectedpar} warps around this function,
where parameter names in \code{nonsetpar} come from is a mystery.
}
\seealso{
Called by \code{\link{geoplot}}, \code{\link{init}} and
\code{\link{selectedpar}}.
}
\keyword{device}
