\name{Elimcomp}
\alias{Elimcomp}

\title{Parameters manipulation.}
\description{
Parameters manipulation.
}
\usage{
Elimcomp(parlist)
}

\arguments{
  \item{parlist}{List of parameters}
}

\value{
Manipulated parameterlist
}

\note{
Needs elaboration, this function is called by \code{\link{geoplot}}, \code{\link{init}} and \code{\link{selectedpar}}.
}

\keyword{manip}
