% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find.hnit.R
\name{find.hnit}
\alias{find.hnit}
\title{Find coordinate(s) ??}
\usage{
find.hnit(pt, poly)
}
\arguments{
\item{pt}{Point(s) ??}

\item{poly}{Polygon (??)}
}
\value{
Returns list with components: \item{x, y }{of coordinates}
}
\description{
Find coordinates with some sort of interpolation (??).
}
\note{
~~further notes~~ Needs elaboration.
}
\seealso{
Called by \code{\link{findcut}}
}
\keyword{manip}

