% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_psw.R
\name{gen_psw}
\alias{gen_psw}
\title{Generate Strong Password}
\usage{
gen_psw(size = 40L, digit = TRUE, lower = TRUE, upper = TRUE, symbol = TRUE)
}
\arguments{
\item{size}{\link[base]{integer} scalar, number of characters allowed.  Default \code{40L}}

\item{digit}{\link[base]{logical} scalar, whether digits \code{0-9} are allowed. Default \code{TRUE}}

\item{lower}{\link[base]{logical} scalar, whether lower case characters \code{a-z} are allowed. Default \code{TRUE}}

\item{upper}{\link[base]{logical} scalar, whether upper case characters \code{A-Z} are allowed. Default \code{TRUE}}

\item{symbol}{\link[base]{logical} scalar, whether commonly used symbols \verb{~!@#$\%^&*<>?:;} are allowed. Default \code{TRUE}}
}
\description{
Generate Strong Password
}
