% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_psw.R
\name{gen_psw}
\alias{gen_psw}
\title{Generate Strong Password}
\usage{
gen_psw(
  size = 40L,
  digit = TRUE,
  lower = TRUE,
  upper = TRUE,
  symbol = character()
)
}
\arguments{
\item{size}{\link[base]{integer} scalar, number of characters allowed.  Default \code{40L}}

\item{digit}{\link[base]{logical} scalar, whether digits \code{0-9} are allowed. Default \code{TRUE}}

\item{lower}{\link[base]{logical} scalar, whether lower case characters \code{a-z} are allowed. Default \code{TRUE}}

\item{upper}{\link[base]{logical} scalar, whether upper case characters \code{A-Z} are allowed. Default \code{TRUE}}

\item{symbol}{\link[base]{character} scalar or \link[base]{vector}, symbols allowed.
Use \code{character()} (default), \code{''} or \code{NULL} to indicate no symbol allowed}
}
\value{
Function \link{gen_psw} returns a \link[base]{character} scalar.
}
\description{
Generate Strong Password
}
\examples{
gen_psw(40L, symbol = '!@#$\%^&*')

}
