% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\docType{package}
\name{geo-package}
\alias{geo-package}
\alias{geo}
\title{Draw and Annotate Maps, Especially Charts of the North Atlantic}
\description{
Used by Hafro staff to draw maps showing the distribution of
fishing intensity and catches, and of survey data 
for Icelandic fish stocks. Potentially useful for others.
}
\author{
Höskuldur Björnsson
}
