% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{geo-package}
\alias{geo}
\alias{geo-package}
\title{geo: Fun Projects with Earth}
\description{
Fun projects with earth based on the function 'plotly::plot_geo()'.
}
\note{
Source of internal datasets,
\describe{
\item{\code{airports_ip2location}}{\url{https://github.com/ip2location/ip2location-iata-icao}}
\item{\code{airports_datahub}}{\url{https://github.com/datasets/airport-codes}}
}
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

Authors:
\itemize{
  \item Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})
}

}
