% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geo.Split.poly.R
\name{geo.Split.poly}
\alias{geo.Split.poly}
\title{Split a polygon into a list}
\usage{
geo.Split.poly(data)
}
\arguments{
\item{data}{A dataframe of coordinates in latitude and longitude}
}
\value{
Returns a list of true polygons as dataframes, each with coordinates
\item{lat}{Latitude} \item{lon}{Longitude}
}
\description{
Split a 'polygon with NAs' into a list of true polygons before calculating
its area.
}
\note{
Needs further elaboration.
}
\seealso{
Called by \code{\link{geoarea}}.
}
\keyword{manip}

