% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geoconvert.1.R
\name{geoconvert.1}
\alias{geoconvert.1}
\title{Convert to decimal degrees}
\usage{
geoconvert.1(x)
}
\arguments{
\item{x}{Vector of decimal-minute-decimal minutes}
}
\value{
Returns converted value in decimal degrees.
}
\description{
Convert to decimal degrees.
}
\seealso{
Called by \code{\link{geoconvert}}
}
\keyword{manip}

