% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geoconvert.2.R
\name{geoconvert.2}
\alias{geoconvert.2}
\title{Convert from decimal degrees}
\usage{
geoconvert.2(lat)
}
\arguments{
\item{lat}{Vector of latitude or longitudes}
}
\value{
Returns a vector of six digit values with degrees, minutes and
fractions of minutes, with two decimal values, concatenated.
}
\description{
Convert from decimal degrees to degrees, minutes and fractional minutes
representation (DDMMmm) of lat or lon.
}
\seealso{
Called by \code{\link{geoconvert}}, when \code{inverse = TRUE}.
}
\keyword{manip}

