% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geodezoom.R
\name{geodezoom}
\alias{geodezoom}
\title{Restores a zoomed plot.}
\usage{
geodezoom()
}
\value{
none
}
\description{
Geodezoom restores a plot zoomed with geodezoom.
}
\section{Side Effects}{
 The limits of the current plot change back to its
original size.
}
\examples{
##    See examples in help(geozoom).
}
\seealso{
\code{\link{geoplot}}, \code{\link{geozoom}}.
}

