% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{geopar}
\alias{geopar}
\title{Parameters for the geo plot functions.}
\format{\preformatted{List of 13
 $ gpar      :List of 57
  ..$ xlog     : logi FALSE
  ..$ ylog     : logi FALSE
  ..$ adj      : num 0.5
  ..$ ann      : logi TRUE
  ..$ ask      : logi FALSE
  ..$ bg       : chr "transparent"
  ..$ bty      : chr "o"
  ..$ cex      : num 0.9
  ..$ cex.axis : num 1
  ..$ cex.lab  : num 1
  ..$ cex.main : num 1.2
  ..$ cex.sub  : num 1
  ..$ col      : chr "black"
  ..$ col.axis : chr "black"
  ..$ col.lab  : chr "black"
  ..$ col.main : chr "black"
  ..$ col.sub  : chr "black"
  ..$ crt      : num 0
  ..$ err      : int -1
  ..$ family   : chr ""
  ..$ fg       : chr "black"
  ..$ font     : int 1
  ..$ font.axis: int 1
  ..$ font.lab : int 1
  ..$ font.main: int 2
  ..$ font.sub : int 1
  ..$ lab      : int [1:3] 5 5 7
  ..$ las      : int 0
  ..$ lend     : chr "round"
  ..$ lheight  : num 1
  ..$ ljoin    : chr "round"
  ..$ lmitre   : num 10
  ..$ lty      : chr "solid"
  ..$ lwd      : num 1
  ..$ mai      : num [1:4] 0.918 0.738 0.738 0.378
  ..$ mar      : num [1:4] 5.1 4.1 4.1 2.1
  ..$ mex      : num 1
  ..$ mgp      : num [1:3] 3 1 0
  ..$ mkh      : num 0.001
  ..$ new      : logi FALSE
  ..$ pch      : int 1
  ..$ pin      : num [1:2] 5.72 5.16
  ..$ plt      : num [1:4] 0.0912 0.9088 0.1314 0.8686
  ..$ ps       : int 18
  ..$ pty      : chr "m"
  ..$ smo      : num 1
  ..$ srt      : num 0
  ..$ tck      : num NA
  ..$ tcl      : num -0.5
  ..$ usr      : num [1:4] -1596 -119 3418 4750
  ..$ xaxp     : num [1:3] -1400 -200 6
  ..$ xaxs     : chr "r"
  ..$ xaxt     : chr "s"
  ..$ xpd      : logi FALSE
  ..$ yaxp     : num [1:3] 3600 4600 5
  ..$ yaxs     : chr "r"
  ..$ yaxt     : chr "s"
 $ limx      : num [1:2] -1541 -173
 $ limy      : num [1:2] 3467 4701
 $ scale     : chr "km"
 $ b0        : num 65
 $ b1        : num 65
 $ l1        : num -18.2
 $ contlab   : num [1:4] 0.117 0.117 0.146 0.883
 $ contlines : num [1:4] 0.117 0.94 0.146 0.883
 $ cont      : logi FALSE
 $ projection: chr "Mercator"
 $ origin    :List of 8
  ..$ lat       : num [1:2] 59.1 70.2
  ..$ lon       : num [1:2] -32.76 -3.68
  ..$ x         : num [1:2] -1541 -173
  ..$ y         : num [1:2] 3467 4701
  ..$ scale     : chr "km"
  ..$ projection: chr "mercator"
  ..$ b0        : num 1.13
  ..$ L0        : NULL
 $ command   : language geoplot(twohmiles)
}}
\usage{
geopar
}
\description{
The \pkg{geo} functions rely on the parameters in \code{options("geopar")}
for plotting. Every time \code{geoplot} is called a new set of geopar is
made and the current erased. If you want to change the parameters in the
background plot after you call geosubplot you have to save the current
geopar before calling geosubplot. When reassigning geopar, the command
\code{options(geopar=list(...))} must be used.
}
\details{
When using \code{geosubplot} it is important to save these parameters before
the subplot is plotted if one wants to make changes to the large plot
afterwards.
}
\examples{
\dontrun{geoplot(type="l")
large.geopar <- geopar                          # Parameters saved.

pos <- list(lat=c(63,64),lon=c(-27,-24))
geosubplot(geoplot(island, new=TRUE,type="l"),pos) # Plot subplot at pos.
geopolygon(island,col=3)                        # Fill with color.
geotext(65,-18,"subplot",col=150)               # Text on subplot.
small.geopar <- geopar                          # Parameters for subplot saved.

geopar <- large.geopar                          # Make big plot active.
geopolygon(island,col=150)                      # Fill with color.
geotext(65,-18.5,"Big plot",csi=8,col=3)        # Text on big plot.

pos <- list(lat=c(63,64),lon=c(-15,-12))
geosubplot(geoplot(island,new=T,type="l"),pos)  # Another subplot drawn.
geopolygon(island,col=3)                        # Fill with color.
geotext(65,-18,"subplot # 2",col=150)           # Text on subplot #2.
small.geopar.2 <- geopar                        # Parameters for subplot saved.

# This must put in a script and sourced together to work like meant to here.
# If done in command line, assign() must be used for changing geopar.
}
}
\keyword{datasets}

