% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoplotpalette.R
\name{geoplotpalette}
\alias{geoplotpalette}
\title{Set define a traditional geoplot-palette}
\usage{
geoplotpalette()
}
\value{
Sets the palette
}
\description{
Set define a traditional geoplot-palette with colors from the matix
'postcol'.
}
\seealso{
\code{\link{colps}}, \code{\link{hsv}}
}
\keyword{~kwd1}
