\name{geoplotpalette}
\alias{geoplotpalette}

\title{
Set define a traditional geoplot-palette
}
\description{
Set define a traditional geoplot-palette with colors from the matix 'postcol'.
}
\usage{
geoplotpalette()
}

\value{
Sets the palette
}

\seealso{
 \code{\link{colps}}, \code{\link{hsv}}
}

\keyword{ ~kwd1 }

