% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_postcodes.R
\name{get_postcode_polygons}
\alias{get_postcode_polygons}
\title{Get postcode polygons}
\usage{
get_postcode_polygons(postcode_list, voronoi, crs)
}
\arguments{
\item{postcode_list}{A list of postcodes}

\item{voronoi}{A simple feature collection containing postcode voronoi}

\item{crs}{Coordinate refrence system for postcodes}
}
\value{
A simple feature collection containing polygons
}
\description{
Get plot data for postcodes from a list of areas
}
