% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IATA.R
\name{IATA}
\alias{IATA}
\title{One or More Air Travel Trips using IATA codes}
\usage{
IATA(x)
}
\arguments{
\item{x}{\link[base]{character} scalar or \link[base]{vector}}
}
\value{
Function \link{IATA} returns an object of S3 class \code{'IATA'}, which is essentially
a \link[base]{list} of \link[base]{integer} \link[base]{vector}s.
}
\description{
One or more air travel trips using International Air Transport Association (IATA) airport codes.
}
\examples{
IATA('NRT-HNL-YVR')
IATA('TFU-LAX-EWR')
IATA('CTU-PEK-ICN-JFK, EWR-IAH-LIM')
IATA('YDF-YVZ')
IATA('TFU-SIN-FRA-JFK')
IATA('IAD-NRT-PVG-TFU')
IATA('PHL-FRA-AAL, AAL-LIS-EWR')
IATA('PEK-ICN-AKL-PPT')
IATA('NRT-HNL-YVR, SEA-YYZ-IAD-VIE-LCA-HKG')
IATA(c('NRT-HNL-YVR', 'SEA-YYZ-IAD-VIE-LCA-HKG')) # same
}
