% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IATA.R
\name{IATA}
\alias{IATA}
\title{One or More Air Travel Trips using IATA codes}
\usage{
IATA(x)
}
\arguments{
\item{x}{\link[base]{character} scalar or \link[base]{vector}}
}
\value{
Function \link{IATA} returns an \code{'IATA'} object, which is essentially
a \link[base]{list} of \link[base]{integer} vectors.
}
\description{
One or more air travel trips using International Air Transport Association (IATA) airport codes.
}
\note{
\code{getMethod(sp::coordinates, signature('SpatialPoints'))}
is slow and will be avoided as much as possible
}
\examples{
IATA('NRT-HNL-YVR')
IATA('TFU-LAX-EWR')
IATA('CTU-PEK-ICN-JFK, EWR-IAH-LIM')
IATA('YDF-YVZ')
IATA('NRT-HNL-YVR, SEA-YYZ-IAD-VIE-LCA-HKG')
IATA(c('NRT-HNL-YVR', 'SEA-YYZ-IAD-VIE-LCA-HKG')) # same

}
