% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra.point.dist.R
\name{intra.point.dist}
\alias{intra.point.dist}
\title{Intra point/position distance}
\usage{
intra.point.dist(x)
}
\arguments{
\item{x}{List of positions with components \code{lat} and \code{lon}.}
}
\value{
Vector of distances between the points in \code{x}.
}
\description{
For a data frame of positions return the vector of intra point distances
}
\examples{

# distances along the perimeter of a statistical rectangle
pos <- rPeri(323)
intra.point.dist(pos)
sum(intra.point.dist(pos))

}
\seealso{
\code{\link{arcdist}}
}
\keyword{arith}
