\name{labels2}
\alias{labels2}

\title{Label plots}

\description{
Label plots with categories.
}

\usage{
labels2(cont, digits, colors, xlim = c(0, 1), ylim = c(0, 1), nx = 4, fill = F)
}

\arguments{
  \item{cont}{Contour ?}
  \item{digits}{Number of digits to use in labels}
  \item{colors}{Colors ?}
  \item{xlim, ylim}{Limits ?}
  \item{nx}{??. Default 4}
  \item{fill}{Fill ?}
}

\value{
No value, labels added to current plot.
}

\note{
Needs elaboration, merge documentation with \code{labels1}, and possibly others?
}

\seealso{
alled by \code{\link{colsymbol}} and \code{\link{geocontour.fill}}.
}

\keyword{aplot}
