% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{leaflet_popup}
\alias{leaflet_popup}
\title{\link[leaflet]{leaflet} with Popup}
\usage{
leaflet_popup(coords, popup = rownames(coords), ...)
}
\arguments{
\item{coords}{2-column \link[base]{matrix} of popup
\emph{longitude} (1st column) and \emph{latitude} (2nd column)}

\item{popup}{\link[base]{character} \link[base]{vector},
popup text.  Default is the \link[base]{rownames} of \code{coords}}

\item{...}{additional parameters, currently not in use}
}
\value{
The function \code{\link[=leaflet_popup]{leaflet_popup()}} returns an \link[leaflet]{leaflet} object, which inherits from \CRANpkg{htmlwidgets}.
}
\description{
..
}
