% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loyalty.R
\docType{class}
\name{loyalty}
\alias{loyalty}
\alias{loyalty-class}
\title{\linkS4class{loyalty} Progress}
\description{
\linkS4class{loyalty} Progress
}
\section{Slots}{

\describe{
\item{\code{program}}{\link[base]{character} scalar,
airline program to which the flight is credited
(\strong{not} carrier of the flight)}

\item{\code{goal}}{\link[base]{numeric} \link[base]{vector},
milestone of each membership tier}

\item{\code{reward}}{\link[base]{numeric} \link[base]{vector},
reward miles earned for each membership tier}

\item{\code{status}}{\link[base]{numeric} \link[base]{vector}
loyalty points earned for each membership tier}

\item{\code{creditcard}}{\link[base]{character} scalar or \link[base]{vector},
co-branded credit card(s)}
}}

\examples{
(JSM24 = new(Class = 'airfare', carrier = 'AA', depart = 'PHL', arrive = 'PDX',
  code = 'J', currency = 'usd', basefare = 276.28, tax = 35.82, upgrade = 241.88))
earnAA(JSM24)
earnAS(JSM24)
earnBA(JSM24)
# earnBAold(JSM24)
}
