% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/paint.window.x.R
\name{paint.window.x}
\alias{paint.window.x}
\title{Paint window for label ?}
\usage{
paint.window.x(listi, col = 0, border = T, poly = T)
}
\arguments{
\item{listi}{Label location ?}

\item{col}{Color, not used ?}

\item{border}{Should border be drawn?}

\item{poly}{Should label be opaque ?}
}
\value{
No value, lines and/or polygon for label added to current geoplot.
}
\description{
Paint window for label based on projected coordinates?.
}
\note{
Needs elaboration and possibly merging with paint.window doc-file.
Argument \code{col} has no effect.
}
\seealso{
Called by \code{\link{geocontour}}.
}
\keyword{aplot}

