% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdistx.R
\name{pdistx}
\alias{pdistx}
\title{Euclidian distance}
\usage{
pdistx(y, x, y1, x1)
}
\arguments{
\item{y, x}{Coordinate vectors of starting positions}

\item{y1, x1}{Coordinate vectors of end positions}
}
\value{
Euclidian distance
}
\description{
Euclidian distance.
}
\note{
Check order of coordinates (?), document distance functions together?
}
\seealso{
Called by \code{\link{variogram}}.
}
\keyword{manip}

