% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IATA.R
\name{plot.IATA}
\alias{plot.IATA}
\title{Draw \code{'IATA'} using \CRANpkg{plotly}}
\usage{
\method{plot}{IATA}(x, ...)
}
\arguments{
\item{x}{\code{'IATA'}}

\item{...}{..}
}
\description{
..
}
\references{
\url{https://plotly.com/r/lines-on-maps/}
}
\keyword{internal}
