% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare.line.R
\name{prepare.line}
\alias{prepare.line}
\title{Prepare a line ?}
\usage{
prepare.line(x)
}
\arguments{
\item{x}{Line vector ?}
}
\value{
Returns a list with components: \item{lx1}{lx1 ?} \item{lx2}{lx2 ?}
\item{nlx}{nlx ?}
}
\description{
Prepare a line for some action ?
}
\note{
Needs elaboration.
}
\seealso{
Called by \code{\link{cut_multipoly}} and \code{\link{findline}}.
}
\keyword{manip}

