% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rlitir.R
\name{Rlitir}
\alias{Rlitir}
\title{Display colors.}
\usage{
Rlitir(n, col)
}
\arguments{
\item{n}{Number of columns and rows of colors to display}

\item{col}{A vector of colors}
}
\value{
No value returned, plots an \code{n} by \code{n} grid of colors.
}
\description{
A grid of colors is plotted.
}
\examples{
# simple, perhaps not so useful application with default palette:

Rlitir(12, 1:144)

# Define a palette with some colors:

ramp <- colorRampPalette(c("khaki1", "gold", "orange",
  "darkorange2", "red", "darkred", "black"))

# number of columns and rows to display

n <- 10

Rlitir(n, ramp(n^2))
}
\seealso{
\code{\link{colorRampPalette}}
}
\keyword{color}

