\name{SegmentWithArrow}
\alias{SegmentWithArrow}

\title{
Plot line segment with arrow at the end.  
}
\description{
Plot line segment with arrow at the end.  
}
\usage{
SegmentWithArrow(pos, angle = 15, size = 0.2, col = "blue", lwd = 2)
}

\arguments{
  \item{pos}{
Segment positions.
}
  \item{angle}{
Angle determining the acuity of the arrow.
}
  \item{size}{
Size of arrow.
}
  \item{col}{
Color of arrow.
}
  \item{lwd}{
Line width of segment.
}
}
\note{
Needs further checking and elaboration.
}

\keyword{aplot}

