% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selectedpar.R
\name{selectedpar}
\alias{selectedpar}
\title{Parameters manipulation ?}
\usage{
selectedpar()
}
\value{
Returns the names of some selected parameters.
}
\description{
Parameters manipulation ?
}
\note{
Needs elaboration,
}
\seealso{
Called by the majority of the graphical functions in geo, calls
\code{\link{Elimcomp}}.
}
\keyword{device}

