\name{spherical}
\alias{spherical}

\title{
Spherical??
}
\description{
'Spherical' of a variogram?
}
\usage{
spherical(rang1, sill, nugget, x)
}

\arguments{
  \item{rang1, sill, nugget}{
Parameters of a variogram.
}
  \item{x}{
The data.
}
}
\value{
Fitted values?
}
\note{
Needs elaboration, if this is a necessary function.
}

\seealso{
Some variogram stuff.
}

\keyword{arith}
